% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.propagation.R
\name{affinity.propagation}
\alias{affinity.propagation}
\title{affinity.propagation}
\usage{
affinity.propagation(idsInSet, score)
}
\arguments{
\item{idsInSet}{A list of set names and their member IDs.}

\item{score}{A vector of addible scores with the same length used to assign input preference;
higher score has larger weight, i.e. -logP.}
}
\value{
A list of \code{clusters} and \code{representatives} for each cluster.
\describe{
 \item{clusters}{A list of character vectors of set IDs in each cluster.}
 \item{representatives}{A character vector of representatives for each cluster.}
}
}
\description{
Use affinity propagation to cluster similar gene sets to reduce redundancy in report.
}
\examples{
\dontrun{
  padj_list = enrichment.results@enrichment.discovery@result$p.adjust
  names(padj_list) = enrichment.results@enrichment.discovery@result$ID

  set_list = enrichment.results@enrichment.discovery@geneSets[names(padj_list)]


  aff.propag <- affinity.propagation(idsInSet = set_list,
                                     score = -log10(padj_list))
}
}
\references{
John M Elizarraras, Yuxing Liao, Zhiao Shi, Qian Zhu, Alexander R Pico, Bing Zhang, WebGestalt 2024: faster gene set analysis and new support for metabolomics and multi-omics, Nucleic Acids Research, Volume 52, Issue W1, 5 July 2024, Pages W415–W421, \url{https://doi.org/10.1093/nar/gkae456}.
}
\author{
Zhiao Shi, Yuxing Liao
}
