% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.external.R
\name{as.MSnSet}
\alias{as.MSnSet}
\title{as.MSnSet}
\usage{
as.MSnSet(DEprot.object, ...)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{...}{Any other parameter accepted by \link{export.external}.}
}
\value{
An \code{MSnSet} object.
}
\description{
Shortcut for \code{export.external(..., format = "MSnSet")}.
}
\examples{
\dontrun{
ms <- as.MSnSet(dpo, counts.type = "imputed")
}

}
\seealso{
\link{export.external}
}
\author{
Sebastian Gregoricchio
}
