% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.external.R
\name{as.QFeatures}
\alias{as.QFeatures}
\title{as.QFeatures}
\usage{
as.QFeatures(DEprot.object, ...)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{...}{Any other parameter accepted by \link{export.external}.}
}
\value{
A \code{QFeatures} object.
}
\description{
Shortcut for \code{export.external(..., format = "QFeatures")}.
}
\examples{
\dontrun{
qf <- as.QFeatures(dpo, assay.name = "proteins")
}

}
\seealso{
\link{export.external}
}
\author{
Sebastian Gregoricchio
}
