% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.external.R
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{as.SummarizedExperiment}
\usage{
as.SummarizedExperiment(DEprot.object, ...)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{...}{Any other parameter accepted by \link{export.external}.}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Shortcut for \code{export.external(..., format = "SummarizedExperiment")}.
}
\examples{
\dontrun{
se <- as.SummarizedExperiment(DEprot::test.toolbox$diff.exp.limma)
}

}
\seealso{
\link{export.external}
}
\author{
Sebastian Gregoricchio
}
