% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.normality.R
\name{check.normality}
\alias{check.normality}
\title{check.normality}
\usage{
check.normality(
  DEprot.object,
  p.threshold = 0.05,
  which.data = "imputed",
  verbose = TRUE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{p.threshold}{Numeric value indicating the p-value threshold to use to define whether a distribution is normal (based on Anderson-Darling normality test). Default: \code{0.05}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{verbose}{Logical value indicating whether messages should be printed. Default: \code{TRUE}.}
}
\value{
A \code{DEprot.normality} object containing: the \code{norm.statement} indicating whether all the samples are normally distributed or not, the \code{norm.AD.tests} list of Anderson-Darling normality test outputs, a list of \code{qqplots} and \code{densities} profiles for each sample.
}
\description{
This function performs principal component analyses (PCA).
}
