% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.pvalues.R
\name{check.pvalues}
\alias{check.pvalues}
\title{check.pvalues}
\usage{
check.pvalues(
  DEprot.analyses.object,
  contrast = 1,
  histogram.binwidth = 0.05,
  p.value.color = "steelblue",
  p.adjusted.color = "indianred"
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Number indicating the position of the contrast to use for the plotting. Default: \code{1}.}

\item{histogram.binwidth}{Numeric value indicating. Default: \code{0.05}}

\item{p.value.color}{String indicating the color to use for the p-values. Default: \code{"steelblue"}.}

\item{p.adjusted.color}{String indicating the color to use for the p-values adjusted. Default: \code{"indianred"}.}
}
\value{
A DEprot.pvalues object containing 3 slots: distribution histogram of p-values and p-values adjusted and, line plot of ranked p-values.
}
\description{
Plots a volcano plot log2(FoldChange) x -log10(p.adjusted) of differential expression results
}
