% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.LFQ.R
\name{contrast.LFQ}
\alias{contrast.LFQ}
\title{contrast.LFQ}
\usage{
contrast.LFQ(
  DEprot.analyses.object,
  contrast = 1,
  show.only.differential = FALSE,
  show.only.significant = FALSE,
  log2FC.positive.color = "firebrick",
  log2FC.negative.color = "steelblue4",
  log2FC.scale.min = NULL,
  log2FC.scale.max = NULL,
  identical.axes = TRUE,
  dot.labels = NULL,
  protein.names.pattern = "",
  labels.in.boxes = FALSE,
  label.font.size = 3,
  label.max.overlaps = 100,
  min.segment.length.labels = 0
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Integer indicating the position of the contrast to use. Default: \code{1}.}

\item{show.only.differential}{Logical value indicating whether only up or down regulated proteins should be plotted. Default: \code{FALSE}.}

\item{show.only.significant}{Logical value indicating whether only proteins with a significant p-value adjusted should be plotted. Notice that this is subordinated to `show.only.differential` (differential proteins pass the P-adj threshold by default). Default: \code{FALSE}.}

\item{log2FC.positive.color}{String indicating any R-supported color for the positive foldchage values. Default: \code{"firebrick"}.}

\item{log2FC.negative.color}{String indicating any R-supported color for the negative foldchage values. Default: \code{"steelblue4"}.}

\item{log2FC.scale.min}{Numeric value indicating the minimum of the foldchange scale (lower values will be colored with the minimum of the scale). Default: \code{NULL}.}

\item{log2FC.scale.max}{Numeric value indicating the maximum of the foldchange scale (lower values will be colored with the maximum of the scale). Default: \code{NULL}.}

\item{identical.axes}{Logical value indicating whether the axes should be forces to be identical between x and y. Default: \code{TRUE}.}

\item{dot.labels}{String vector indicating labels to show on the plot that should correspond to \code{prot.id} column values. Default: \code{NULL} (no labels shown).}

\item{protein.names.pattern}{Character indicating a regular expression to remove from the protein IDs. Default: \code{""}, no alterations in the protein IDs.}

\item{labels.in.boxes}{Logical value indicating whether the labels should be visualized as boxes. Default: \code{FALSE}.}

\item{label.font.size}{Numeric value indicating the size to use for the dot labels. Default: \code{3}.}

\item{label.max.overlaps}{Numeric value indicating the maximum number of overlaps allowed between labels. Default: \code{100}.}

\item{min.segment.length.labels}{Numeric value indicating the minimal length of the segments that connect the labels to the points. Default: \code{0} (segment always shown).}
}
\value{
A scatter plot of class ggplot2.
}
\description{
Plots a scatter of the average LFQ values of each condition in a contrast. Dots will be colored by FoldChange and the size will correspond to the -log10(Padj).
}
\examples{
contrast.LFQ(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
             contrast = 1)


}
