% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.scatter.R
\name{contrast.scatter}
\alias{contrast.scatter}
\title{contrast.scatter}
\usage{
contrast.scatter(
  DEprot.analyses.object,
  contrast.x,
  contrast.y,
  regression.line.color = "firebrick",
  correlation.method = "pearson",
  add.foldchange.threshold = TRUE,
  symmetric.axes = TRUE
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast.x}{Integer indicating the position of the contrast to use for the x-axis of the plot.}

\item{contrast.y}{Integer indicating the position of the contrast to use for the y-axis of the plot.}

\item{regression.line.color}{String indicating any R-supported color to use for the regression line and error. Default: \code{"firebrick"}.}

\item{correlation.method}{String indicating the clustering method to use to generate the correlation matrix. Possible options: 'pearson', 'spearman', 'kendall'. Default: \code{"pearson"}.}

\item{add.foldchange.threshold}{Logical value to indicate whether two gray rectangles should be used to highlight the non-differential area in the plot, based on the foldchange threshold indicated during the differential analyses. Default: \code{TRUE}.}

\item{symmetric.axes}{Logical value indicating whether the axes should be forces to be symmetric between x and y. Default: \code{TRUE}.}
}
\value{
A scatter plot of class ggplot2.
}
\description{
Plots a scatter plot of the log2(fold change expression) derived from two differential analyses contrasts.
}
\examples{
contrast.scatter(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
                 contrast.x = 1,
                 contrast.y = 2) +
  ggplot2::ylab("log~2~(Fold Change 6h.10nM.E2/6h.DMSO)") +
  ggplot2::xlab("log~2~(Fold Change FBS/6h.DMSO)")


}
