% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum_v5.0}
\alias{corum_v5.0}
\title{CORUM version 5.0}
\format{
A data frame with 24705 rows and 4 columns:
\describe{
  \item{\code{complex.id}}{Numeric value idnciating the complex ID}
  \item{\code{complex.name}}{Extended name of the complex}
  \item{\code{organism}}{Specie to which the complex belong}
  \item{\code{protein.members}}{List of the proteins}
}
}
\source{
https://mips.helmholtz-muenchen.de/corum/download
}
\usage{
corum_v5.0
}
\description{
Collection of CORUM complexes for all organisms available.
}
\keyword{datasets}
