% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{DEprot-options}
\alias{DEprot-options}
\title{Package options}
\description{
The \code{DEprot} package uses the following global options
(set via \code{\link[base:options]{options()}}):

\describe{
  \item{\code{DEprot.update_check}}{
    Logical. Whether DEprot checks for package updates.
    Default is \code{TRUE}.
  }
}
}
\keyword{internal}
