% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.analyses-class}
\alias{DEprot.analyses-class}
\title{DEprot.analyses class}
\description{
Extension of the \linkS4class{DEprot} class: it stores the same data and
  pre-processing information, plus the results of the differential analyses.
  All the slots of \linkS4class{DEprot} are inherited and behave identically.
}
\section{Slots}{

\describe{
\item{\code{analyses.result.list}}{List containing the differential results for each contrast. Class: \code{"ANY"}. The list contains the following elements:
\describe{
  \item{\code{results}: }{a data.frame containing the results of the analyses; includes average expression of each group, basemean, foldchange, pvalue and p.adj, test statistic, degrees of freedom and differential.status}
  \item{\code{n.diff}: }{a summary table showing the number of proteins in each differential expression status (up/down/unresponsive, null)}
  \item{\code{PCA.data}: }{output of \link{perform.PCA} for the subset of samples analyzed in a specific contrast}
  \item{\code{PCA.plots}: }{combination of 3 plots: scatter PC1-vs-PC2, scatter PC2-vs-PC3, and cumulative bar plot}
  \item{\code{correlations}: }{combination of Pearson and Spearman correlation heatmaps (obtained by \link{plot.correlation.heatmap}) for the subset of samples analyzed in a specific contrast}
  \item{\code{volcano}: }{volcano plot showing the log2(FoldChange) x -log10(p.adjusted) of differential expression results; it can be regenerated using \link{plot.volcano}}
  \item{\code{MA.plot}: }{MA-plot showing the log2(basemean) x log2(FoldChange) of differential expression results; it can be regenerated using \link{plot.MA}}}}

\item{\code{contrasts}}{List of contrasts. each contrast is a vector indicating, in the order: metadata.table.column - groupA - groupB; (groupA / group B). Class: \code{"ANY"}.}

\item{\code{differential.analyses.params}}{List of parameters used to run the differential analyses (fold change thresholds, p-value threshold, p-adjustment method, etc.). Class: \code{"ANY"}.}
}}

