% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.contrast.heatmap}
\alias{DEprot.contrast.heatmap}
\title{DEprot.contrast.heatmap class}
\description{
DEprot.contrast.heatmap class
}
\section{Slots}{

 \describe{
   \item{\code{heatmap}:}{Ggplot object corresponding to any heatmap generate by either \link{heatmap.contrasts}. Class: \code{"ANY"}.}
   \item{\code{cluster}:}{The \code{hclust} object of the rows (proteins). Class: \code{"ANY"}.}
 }
}

