% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.correlation}
\alias{DEprot.correlation}
\title{DEprot.correlation class}
\description{
DEprot.correlation class
}
\section{Slots}{

 \describe{
   \item{\code{heatmap}:}{Ggplot object corresponding to the correlation heatmap. Class: \code{"ANY"}.}
   \item{\code{corr.metadata}:}{metadata of the samples used in the correlation (subset of the original \code{DEprot@metadata}. Class: \code{"ANY"}.}
   \item{\code{sample.subset}:}{vector containing the list of samples analyzed. Class: \code{"ANY"}.}
   \item{\code{data.used}:}{vector indicating the type of counts used (imputed, normalized, raw). Class: \code{"ANY"}.}
   \item{\code{corr.matrix}:}{the correlation matrix on which the heatmap is base on. Class: \code{"ANY"}.}
   \item{\code{distance}:}{object of class \code{dist} corresponding to the output of \code{as.dist(1 - correlation.matrix)}. Class: \code{"ANY"}.}
   \item{\code{cluster}:}{\code{hclust} object generated by \code{hclust(d = as.dist(1 - correlation.matrix), method = clustering.method)}. Class: \code{"ANY"}.}
 }
}

