% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.counts.heatmap}
\alias{DEprot.counts.heatmap}
\title{DEprot.counts.heatmap class}
\description{
DEprot.counts.heatmap class
}
\section{Slots}{

\describe{
\item{\code{heatmap}}{Ggplot object corresponding to any heatmap generate by either \link{heatmap.counts}. Class: \code{"ANY"}.}

\item{\code{row.cluster}}{The \code{hclust} object of the rows (proteins). Class: \code{"ANY"}.}

\item{\code{column.cluster}}{The \code{hclust} object of the columns (samples). Class: \code{"ANY"}.}
}}

