% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.PCoA-class}
\alias{DEprot.PCoA-class}
\title{DEprot.PCoA class}
\description{
DEprot.PCoA class
}
\section{Slots}{

\describe{
\item{\code{PCoA.metadata}}{metadata of the samples used in the PCoA (subset of the original \code{DEprot@metadata}). Class: \code{"ANY"}.}

\item{\code{sample.subset}}{vector containing the list of samples analyzed. Class: \code{"ANY"}.}

\item{\code{data.used}}{vector indicating the type of counts used (imputed, normalized, raw), inherited from the \code{DEprot.correlation} object. Class: \code{"ANY"}.}

\item{\code{correlation.method}}{string indicating the correlation method used to build the dissimilarities ('pearson', 'spearman', 'kendall'). Class: \code{"ANY"}.}

\item{\code{correlation.matrix}}{the sample-by-sample correlation matrix the PCoA is based on. Class: \code{"ANY"}.}

\item{\code{distance.input}}{object of class \code{dist} as stored in the input \code{DEprot.correlation} object, i.e. \code{as.dist(1 - correlation.matrix)}. Class: \code{"ANY"}.}

\item{\code{distance}}{object of class \code{dist} effectively used for the ordination, after transformation and/or non-euclidean correction. Class: \code{"ANY"}.}

\item{\code{distance.method}}{string describing how the dissimilarities used were obtained. Class: \code{"ANY"}.}

\item{\code{cmdscale}}{full output of \code{stats::cmdscale} (\code{points}, \code{eig}, \code{GOF}, \code{ac}). Class: \code{"ANY"}.}

\item{\code{PCos}}{data.frame combining the principal coordinates and the metadata table, useful for replotting. Class: \code{"ANY"}.}

\item{\code{eigenvalues}}{numeric vector of all the eigenvalues returned by the ordination. Class: \code{"ANY"}.}

\item{\code{importance}}{statistical summary table of the ordination per each principal coordinate (eigenvalue, proportion/percentage of variance, cumulative proportion, broken-stick expectation). Class: \code{"ANY"}.}

\item{\code{euclidean.diagnostics}}{list summarizing the amount and the magnitude of negative eigenvalues, and the goodness-of-fit of the embedding. Class: \code{"ANY"}.}

\item{\code{axis.loadings}}{data.frame of protein-vs-axis correlations (only if a \code{DEprot} object was provided to \link{perform.PCoA}, otherwise \code{NULL}). Class: \code{"ANY"}.}

\item{\code{scatter.plot}}{ggplot object corresponding to the scatter of the two principal coordinates requested by the user. Class: \code{"ANY"}.}

\item{\code{scatter.plot.123}}{patchwork object combining the PCo1-vs-PCo2 and PCo3-vs-PCo2 scatters. Class: \code{"ANY"}.}

\item{\code{cumulative.PCo.plot}}{ggplot object showing the proportion of variance and its cumulative curve for each principal coordinate. Class: \code{"ANY"}.}

\item{\code{shepard.plot}}{ggplot object comparing the input dissimilarities with the distances reconstructed in the reduced space (embedding quality diagnostic). Class: \code{"ANY"}.}

\item{\code{parameters}}{list of the parameters used to compute the ordination. Class: \code{"ANY"}.}
}}

