% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.power-class}
\alias{DEprot.power-class}
\title{DEprot.power class}
\description{
DEprot.power class
}
\section{Slots}{

\describe{
\item{\code{power.table}}{Data.frame reporting, for each sample size tested: the per-test significance level required to control the FDR, the average power, the expected number of true and false discoveries and the expected FDR. Class: \code{"ANY"}.}

\item{\code{n.required}}{Numeric value indicating the smallest number of samples per group reaching the target average power (\code{NA} when the target is not reached within the range tested). Class: \code{"ANY"}.}

\item{\code{effect.size}}{Numeric vector of the standardized effect sizes (|d|) used for the estimation. Class: \code{"ANY"}.}

\item{\code{pi0}}{Numeric value indicating the proportion of non-differential proteins used for the estimation. Class: \code{"ANY"}.}

\item{\code{power.plot}}{Ggplot object showing the average power as function of the number of samples per group. Class: \code{"ANY"}.}

\item{\code{discoveries.plot}}{Ggplot object showing the expected number of true discoveries as function of the number of samples per group. Class: \code{"ANY"}.}

\item{\code{effect.size.plot}}{Ggplot object showing the distribution of the standardized effect sizes used. Class: \code{"ANY"}.}

\item{\code{params}}{List of the parameters used for the estimation. Class: \code{"ANY"}.}
}}

