% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot}
\alias{DEprot}
\title{DEprot class}
\description{
DEprot class
}
\section{Slots}{

 \describe{
   \item{\code{metadata}:}{The data.frame corresponding to the metadata table describing the samples. Class: \code{"ANY"}.}
   \item{\code{raw.counts}:}{Numeric matrix (rows: proteins, columns: samples) of the raw counts. Class: \code{"ANY"}.}
   \item{\code{norm.counts}:}{Numeric matrix (rows: proteins, columns: samples) of the normalized counts. Class: \code{"ANY"}.}
   \item{\code{imputed.counts}:}{Numeric matrix (rows: proteins, columns: samples) of the imputed counts. Class: \code{"ANY"}.}
   \item{\code{log.base}:}{Numeric value indicating the base of the logarithm expressing the values in the loaded data. Class: \code{"ANY"}.}
   \item{\code{log.transformed}:}{Logical value indicating whether the data are log-transformed or not. Class: \code{"logical"}.}
   \item{\code{imputed}:}{Logical value indicating whether the data are imputed. Class: \code{"logical"}.}
   \item{\code{imputation}:}{String (or any other class) value indicating the imputation method. Class: \code{"ANY"}.}
   \item{\code{normalized}:}{Logical value indicating whether the data are normalized. Class: \code{"logical"}.}
   \item{\code{normalization.method}:}{String (or any other class) value indicating the normalization method. Class: \code{"ANY"}. Class: \code{"ANY"}.}
   \item{\code{boxplot.raw}:}{Ggplot object showing the distribution of the raw values per sample. Class: \code{"ANY"}.}
   \item{\code{boxplot.norm}:}{Ggplot object showing the distribution of the normalized values per sample. Class: \code{"ANY"}.}
   \item{\code{boxplot.imputed}:}{Ggplot object showing the distribution of the imputed values per sample. Class: \code{"ANY"}.}
   \item{\code{analyses.result.list}:}{For this type of object the value is \code{NULL}. Class: \code{"ANY"}.}
   \item{\code{contrasts}:}{For this type of object the value is \code{NULL}. Class: \code{"ANY"}.}
   \item{\code{differential.analyses.params}:}{For this type of object the value is \code{NULL}. Class: \code{"ANY"}.}
 }
}

