% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\docType{class}
\name{DEprot.RMSE-class}
\alias{DEprot.RMSE-class}
\title{DEprot.RMSE class}
\description{
DEprot.RMSE class
}
\section{Slots}{

\describe{
\item{\code{original.DEprot.object}}{Object of class \code{DEprot} used to compute the RMSE. Class: \code{"ANY"}.}

\item{\code{percentage.test}}{Percentage of the total proteins that should be used to perform the comparisons. Class: \code{"ANY"}.}

\item{\code{seed}}{Seed used for the randomization. Class: \code{"ANY"}.}

\item{\code{fraction.missing.values}}{Fraction of missing values in the original table. Class: \code{"ANY"}.}

\item{\code{test.dataset}}{Subset of the original table used for the comparisons. Class: \code{"ANY"}.}

\item{\code{imputed.objects}}{List of the output of \link{impute.counts} (class \code{DEprot}) using the different imputation methods. Class: \code{"ANY"}.}

\item{\code{RMSE.tables}}{List of data.frames, one per tested imputation method, containing the coordinates of the imputed values and including the following columns (Class: \code{"ANY"}.):
\describe{
  \item{\code{row.id}: }{id of the row (protein)}
  \item{\code{col.id}: }{id of the column (sample)}
  \item{\code{expected.values}: }{value measured in the experiment}
  \item{\code{imputation.method}: }{id of the method used for the imputation}
  \item{\code{imputed.values}: }{value imputed by \code{DEprot}}
  \item{\code{residuals}: }{difference of the the values, \code{imputed - expected}}
  \item{\code{sq.residuals}: }{the squared value of the residuals}}}

\item{\code{RMSE.scores}}{A table indicating the methods and the corresponding RMSE score. Class: \code{"ANY"}.}

\item{\code{correlation.plots}}{A list of ggplot objects with the correlation between observed/expected and imputed values. Class: \code{"ANY"}.}

\item{\code{density.residuals}}{A list of ggplot objects depicting the distribution of the residuals. Class: \code{"ANY"}.}

\item{\code{masking}}{List describing how the values of the test dataset have been hidden, containing the following elements (Class: \code{"ANY"}.):
\describe{
  \item{\code{mode}: }{masking actually applied, \code{"MCAR"} or \code{"intensity"}. It reports \code{"MCAR"} also when the intensity-dependent masking was required but none of the dropout curves could be estimated}
  \item{\code{dropout.by}: }{level at which the dropout curves have been estimated, \code{"group"} or \code{"sample"}}
  \item{\code{dropout.curves}: }{list with the intercept and the slope of the logistic curve estimated for each group of replicates (or each sample), \code{NULL} when the masking is \code{"MCAR"}}
  \item{\code{max.NA.per.row}: }{maximum number of missing values that a protein could already have to enter the test dataset}
  \item{\code{max.NA.fraction}: }{maximum fraction of missing values that a protein could reach after the masking}}}
}}

