% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\docType{class}
\name{DEprot.timecourse.enrichment-class}
\alias{DEprot.timecourse.enrichment-class}
\title{DEprot.timecourse.enrichment class}
\description{
DEprot.timecourse.enrichment class
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame combining the enrichment results of all the clusters, with a 'cluster' column. Class: \code{"ANY"}.}

\item{\code{enrichment.per.cluster}}{named list containing the \code{enrichResult} object returned by \code{clusterProfiler::enricher} for each cluster. Class: \code{"ANY"}.}

\item{\code{dotplot}}{ggplot object showing, for each cluster, the enriched genesets; the number inside each dot corresponds to the count of proteins. Class: \code{"ANY"}.}

\item{\code{universe}}{vector containing the background gene list used for the tests. Class: \code{"ANY"}.}

\item{\code{parameters}}{a list containing the parameters used to run the analyses. Class: \code{"ANY"}.}
}}

