% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.analyses.R
\name{diff.analyses}
\alias{diff.analyses}
\title{diff.analyses}
\usage{
\method{diff}{analyses}(
  DEprot.object,
  contrast.list,
  replicate.column = NULL,
  linear.FC.th = 2,
  linear.FC.unresp.range = c(1/1.1, 1.1),
  padj.th = 0.05,
  padj.method = "BH",
  stat.test = "t.test",
  paired.test = FALSE,
  up.color = "indianred",
  down.color = "steelblue",
  unresponsive.color = "purple",
  null.color = "gray",
  which.data = "imputed",
  overwrite.analyses = FALSE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{contrast.list}{List of 3-elements vectors indicating (in order): metadata_column, variable_1, variable_2.}

\item{replicate.column}{String indicating the name of a column from the metadata table in which are stored the replicate IDs. This column is used only if \code{paired.test = TRUE}. Default: \code{NULL}.}

\item{linear.FC.th}{Number indicating the (absolute) fold change threshold (linear scale) to use to define differential proteins. Default: \code{2}.}

\item{linear.FC.unresp.range}{A numeric 2-elements vector indicating the range (linear scale) used to define the unresponsive fold changes. Default: \code{c(1/1.1, 1.1)}.}

\item{padj.th}{Numeric value indicating the p.adjusted threshold to apply to the differential analyses. Default: \code{0.05}.}

\item{padj.method}{String indicating the method to use to correct the p-values. One among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: \code{BH}.}

\item{stat.test}{String indicating the type of statistic test to use. One among: "t-test" and "wilcoxon". Default: \code{"t.test"}.}

\item{paired.test}{Logical value indicating whether paired statistical test should be performed. For each contrast it will be verified that replicate IDs are not repeated within a group and, that replicate number and the identifiers coincide between the two groups. Default: \code{FALSE}.}

\item{up.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"indianred"}.}

\item{down.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"steelblue"}.}

\item{unresponsive.color}{String indicating the color to use for unresponsive proteins in the plots. Default: \code{"purple"}.}

\item{null.color}{String indicating the color to use for null proteins in the plots. Default: \code{"gray"}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{overwrite.analyses}{Logical value to indicate whether overwrite analyses already generated. Default: \code{FALSE}.}
}
\value{
An object of class \code{DEprot.analyses}
}
\description{
Allows for the computation of differential analyses using multiple t/Wilcox tests. Includes means, Fold Changes, and p-values.
}
\examples{
# Basic Differential analyses with paired analyses
dpo <- diff.analyses(DEprot.object = DEprot::test.toolbox$dpo.imp,
                     contrast.list = list(c("condition", "FBS", "6h.DMSO"),
                                          c("condition", "6h.10nM.E2", "6h.DMSO")),
                     linear.FC.th = 1.5)


# Differential analyses with paired analyses
dpo <- diff.analyses(DEprot.object = DEprot::test.toolbox$dpo.imp,
                     contrast.list = list(c("condition", "FBS", "6h.DMSO"),
                                          c("condition", "6h.10nM.E2", "6h.DMSO")),
                     paired.test = TRUE,
                     replicate.column = "replicate",
                     linear.FC.th = 1.5)


}
\author{
Sebastian Gregoricchio
}
