% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.protein.info.R
\name{.append.protein.info}
\alias{.append.protein.info}
\title{.append.protein.info}
\usage{
.append.protein.info(
  data,
  DEprot.object = NULL,
  protein.info.columns = "none",
  protein.info.prefix = NULL
)
}
\arguments{
\item{data}{data.frame containing a \code{prot.id} column.}

\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}, or \code{NULL}.}

\item{protein.info.columns}{String vector of the columns to append, or one of the keywords \code{"none"} and \code{"all"}.}

\item{protein.info.prefix}{String to prepend to the names of the appended columns, or \code{NULL}.}
}
\value{
The input data.frame, with the annotation columns appended when available.
}
\description{
Internal. Appends the columns of the \code{protein.info} table to a results table,
matching the rows by protein ID rather than by position, since the results might have been
re-ordered or filtered.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
