% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.export.R
\name{.build.qfeatures}
\alias{.build.qfeatures}
\title{.build.qfeatures}
\usage{
.build.qfeatures(
  assay.list,
  row.data,
  col.data,
  obj.metadata,
  assay.name = "proteins",
  install.missing = "ask"
)
}
\arguments{
\item{assay.list}{Named list of matrices sharing the same dimnames.}

\item{row.data}{A data.frame.}

\item{col.data}{A data.frame.}

\item{obj.metadata}{A named list.}

\item{assay.name}{String, name of the assay set.}

\item{install.missing}{One among \code{"ask"}, \code{"always"}, \code{"never"}.}
}
\value{
A \code{QFeatures} object.
}
\description{
Internal helper building a \code{QFeatures} object containing a single, protein-level
  assay set.
}
\keyword{internal}
