% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.export.R
\name{.build.summarized.experiment}
\alias{.build.summarized.experiment}
\title{.build.summarized.experiment}
\usage{
.build.summarized.experiment(
  assay.list,
  row.data,
  col.data,
  obj.metadata,
  install.missing = "ask"
)
}
\arguments{
\item{assay.list}{Named list of matrices sharing the same dimnames.}

\item{row.data}{A data.frame.}

\item{col.data}{A data.frame.}

\item{obj.metadata}{A named list.}

\item{install.missing}{One among \code{"ask"}, \code{"always"}, \code{"never"}.}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Internal helper building a \code{SummarizedExperiment} from the collected assays and
  annotations.
}
\keyword{internal}
