% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.filter.long.report}
\alias{.filter.long.report}
\title{.filter.long.report}
\usage{
.filter.long.report(
  tb,
  id.col,
  precursor.q.col = NA,
  protein.q.col = NA,
  q.value = 0.01,
  pg.q.value = 0.01,
  remove.contaminants = TRUE,
  contaminant.pattern = NULL,
  verbose = TRUE
)
}
\arguments{
\item{tb}{A data.frame.}

\item{id.col}{String, protein ID column.}

\item{precursor.q.col}{String (or NA), column of the precursor q-values.}

\item{protein.q.col}{String (or NA), column of the protein-group q-values.}

\item{q.value}{Numeric or NULL.}

\item{pg.q.value}{Numeric or NULL.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}

\item{verbose}{Logical.}
}
\value{
A data.frame.
}
\description{
Internal helper applying q-value and contaminant filters to a long report.
}
\keyword{internal}
