% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.protein.info.R
\name{.get.protein.ids}
\alias{.get.protein.ids}
\title{.get.protein.ids}
\usage{
.get.protein.ids(object)
}
\arguments{
\item{object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}
}
\value{
A character vector, or \code{NULL} if no counts table is available.
}
\description{
Extracts the protein IDs (row names) from the first available
counts table of a \code{DEprot}/\code{DEprot.analyses} object. All the counts
tables of an object share the same rows, hence the first available one is
representative of the whole object.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
