% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.protein.info.R
\name{.get.protein.info}
\alias{.get.protein.info}
\title{.get.protein.info}
\usage{
.get.protein.info(object)
}
\arguments{
\item{object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}
}
\value{
A \code{data.frame} or \code{NULL}.
}
\description{
Safe accessor for the \code{protein.info} slot. Returns \code{NULL}
both when the slot is empty and when the object was created with a version of
\code{DEprot} predating the introduction of the slot (backward compatibility).
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
