% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.PCoA.R
\name{.lingoes.correction}
\alias{.lingoes.correction}
\title{.lingoes.correction}
\usage{
.lingoes.correction(d)
}
\arguments{
\item{d}{An object of class \code{dist}.}
}
\value{
An object of class \code{dist}.
}
\description{
Internal. Applies the Lingoes (1971) correction to a \code{dist} object, i.e.
\code{d_prime = sqrt(d^2 + 2*c)} for \code{i != j}, where \code{c} is the absolute value of the most
negative eigenvalue of the Gower-centered matrix. This makes the dissimilarities euclidean-embeddable.
}
\keyword{internal}
