% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.enrichment.R
\name{.parse.ratio}
\alias{.parse.ratio}
\title{.parse.ratio}
\usage{
.parse.ratio(x, position = "ratio")
}
\arguments{
\item{x}{Vector of strings in the form 'x/y'.}

\item{position}{String indicating which value to return: \code{"ratio"} (x/y), \code{"numerator"} (x) or \code{"denominator"} (y). Default: \code{"ratio"}.}
}
\value{
A numeric vector.
}
\description{
Internal. Converts the 'x/y' strings returned by clusterProfiler into a numeric ratio.
}
\keyword{internal}
