% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.PCoA.R
\name{.PCoA.axis.loadings}
\alias{.PCoA.axis.loadings}
\title{.PCoA.axis.loadings}
\usage{
.PCoA.axis.loadings(DEprot.object, which.data, coordinates, verbose = TRUE)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{which.data}{String indicating the count type to use.}

\item{coordinates}{data.frame of the principal coordinates (samples as rows).}

\item{verbose}{Logical value indicating whether messages should be printed.}
}
\value{
A data.frame of protein-vs-axis correlations.
}
\description{
Internal. Computes the a-posteriori projection of the proteins onto the principal
coordinates, as the correlation between each protein profile and each axis.
}
\keyword{internal}
