% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.diann.matrix}
\alias{.read.diann.matrix}
\title{.read.diann.matrix}
\usage{
.read.diann.matrix(tb, protein.id, remove.contaminants, contaminant.pattern)
}
\arguments{
\item{tb}{A data.frame.}

\item{protein.id}{String or \code{"auto"}.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}
}
\value{
A list.
}
\description{
Internal reader for the DIA-NN \code{*_matrix.tsv} files
  (pg_matrix, gg_matrix, pr_matrix, unique_genes_matrix).
}
\keyword{internal}
