% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.fragpipe}
\alias{.read.fragpipe}
\title{.read.fragpipe}
\usage{
.read.fragpipe(
  tb,
  quantity,
  protein.id,
  remove.contaminants,
  contaminant.pattern
)
}
\arguments{
\item{tb}{A data.frame.}

\item{quantity}{String or \code{"auto"}.}

\item{protein.id}{String or \code{"auto"}.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}
}
\value{
A list.
}
\description{
Internal reader for FragPipe \code{combined_protein.tsv} / \code{combined_peptide.tsv}.
}
\keyword{internal}
