% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.generic}
\alias{.read.generic}
\title{.read.generic}
\usage{
.read.generic(tb, id.column, sample.columns)
}
\arguments{
\item{tb}{A data.frame.}

\item{id.column}{String, protein ID column.}

\item{sample.columns}{Character vector (optionally named) of the intensity columns.}
}
\value{
A list.
}
\description{
Internal reader for any wide table, when the user provides the ID and quantity columns.
}
\keyword{internal}
