% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.maxquant}
\alias{.read.maxquant}
\title{.read.maxquant}
\usage{
.read.maxquant(
  tb,
  quantity,
  protein.id,
  remove.contaminants,
  contaminant.pattern,
  verbose
)
}
\arguments{
\item{tb}{A data.frame.}

\item{quantity}{String or \code{"auto"}.}

\item{protein.id}{String or \code{"auto"}.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}

\item{verbose}{Logical.}
}
\value{
A list.
}
\description{
Internal reader for MaxQuant \code{proteinGroups.txt}.
}
\keyword{internal}
