% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.msstats.R
\name{.read.msstats.lfq}
\alias{.read.msstats.lfq}
\title{.read.msstats.lfq}
\usage{
.read.msstats.lfq(
  tb,
  protein.id,
  sample.id,
  sample.sep,
  annotation.columns,
  column.id,
  log.base
)
}
\arguments{
\item{tb}{A data.frame.}

\item{protein.id}{String or \code{"auto"}.}

\item{sample.id}{Character vector or \code{"auto"}.}

\item{sample.sep}{String.}

\item{annotation.columns}{Character vector or \code{"auto"}.}

\item{column.id}{String.}

\item{log.base}{Numeric.}
}
\value{
A list.
}
\description{
Internal reader for the protein-level output of \code{MSstats::dataProcess()}.
}
\keyword{internal}
