% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.proteome.discoverer}
\alias{.read.proteome.discoverer}
\title{.read.proteome.discoverer}
\usage{
.read.proteome.discoverer(
  tb,
  quantity,
  protein.id,
  remove.contaminants,
  contaminant.pattern
)
}
\arguments{
\item{tb}{A data.frame.}

\item{quantity}{String or \code{"auto"}.}

\item{protein.id}{String or \code{"auto"}.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}
}
\value{
A list.
}
\description{
Internal reader for Proteome Discoverer protein-level exports
  (R-friendly headers recommended).
}
\keyword{internal}
