% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.spectronaut.pivot}
\alias{.read.spectronaut.pivot}
\title{.read.spectronaut.pivot}
\usage{
.read.spectronaut.pivot(
  tb,
  quantity,
  protein.id,
  remove.contaminants,
  contaminant.pattern
)
}
\arguments{
\item{tb}{A data.frame.}

\item{quantity}{String or \code{"auto"}.}

\item{protein.id}{String or \code{"auto"}.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}
}
\value{
A list.
}
\description{
Internal reader for Spectronaut pivot exports, i.e. one column per run named
  \code{"[1] sample.raw.PG.Quantity"}.
}
\keyword{internal}
