% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.external.R
\name{.read.spectronaut}
\alias{.read.spectronaut}
\title{.read.spectronaut}
\usage{
.read.spectronaut(
  tb,
  quantity,
  protein.id,
  sample.id,
  summarization,
  q.value,
  pg.q.value,
  remove.contaminants,
  contaminant.pattern,
  install.missing,
  verbose
)
}
\arguments{
\item{tb}{A data.frame.}

\item{quantity}{String or \code{"auto"}.}

\item{protein.id}{String or \code{"auto"}.}

\item{sample.id}{String or \code{"auto"}.}

\item{summarization}{One among \code{"maxlfq"}, \code{"sum"}, \code{"median"}, \code{"none"}.}

\item{q.value}{Numeric or NULL.}

\item{pg.q.value}{Numeric or NULL.}

\item{remove.contaminants}{Logical.}

\item{contaminant.pattern}{String or NULL.}

\item{install.missing}{One among \code{"ask"}, \code{"always"}, \code{"never"}.}

\item{verbose}{Logical.}
}
\value{
A list.
}
\description{
Internal reader for long-format Spectronaut reports.
}
\keyword{internal}
