% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{.splsda.axis.labels}
\alias{.splsda.axis.labels}
\title{.splsda.axis.labels}
\usage{
.splsda.axis.labels(importance, axis)
}
\arguments{
\item{importance}{importance table of the \code{DEprot.sPLSDA} object.}

\item{axis}{Number of the component.}
}
\value{
A string.
}
\description{
Internal. Builds the axis titles of the sPLS-DA scatters. The generic ordination label ('comp1 (12.3%)') is replaced by an explicit one, since on a supervised model the percentage refers to the protein block and not to the separation between the classes.
}
\keyword{internal}
