% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{.tc.check.values}
\alias{.tc.check.values}
\title{.tc.check.values}
\usage{
.tc.check.values(values)
}
\arguments{
\item{values}{String indicating the quantity to plot. One among: 'zscore', 'log2FC', 'counts' (aliases accepted).}
}
\value{
One among "zscore", "log2FC", "counts".
}
\description{
Normalizes the \code{values} argument of the time-course plots, accepting aliases.
}
\keyword{internal}
