% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{.tc.transform.time}
\alias{.tc.transform.time}
\title{.tc.transform.time}
\usage{
.tc.transform.time(time, time.transform)
}
\arguments{
\item{time}{Numeric vector of the timepoints.}

\item{time.transform}{String indicating the transformation. One among: 'none', 'log2', 'log10', 'log1p', 'log', 'ln', 'sqrt'.}
}
\value{
A numeric vector of the transformed timepoints.
}
\description{
Applies a transformation to the numeric time vector. The log transformations are computed as log(t + 1) so that time zero is allowed.
}
\keyword{internal}
