% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{.tc.zscore}
\alias{.tc.zscore}
\title{.tc.zscore}
\usage{
.tc.zscore(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
A matrix of the same dimensions.
}
\description{
Row-wise Z-score of a matrix. The rows with a null variance are returned as zeros instead of NaN.
}
\keyword{internal}
