% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpo.imputed.counts}
\alias{dpo.imputed.counts}
\title{Imputed proteomics (LFQ) data}
\format{
Object of class \code{DEprot}
}
\source{
Simulated data
}
\usage{
data("dpo.imputed.counts")
}
\description{
Dummy example of full proteomics data (LFQ values in log2). Imputed using missForest algorithm.
}
\keyword{datasets}
