% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.proteins.R
\name{filter.proteins}
\alias{filter.proteins}
\title{filter.proteins}
\usage{
\method{filter}{proteins}(DEprot.object, proteins, mode = "keep")
}
\arguments{
\item{DEprot.object}{Object of class \code{DEprot}.}

\item{proteins}{String vector indicating a list of proteins to be use as filtering criteria. It must correspond to the counts row names and 'prot.id' column of the analyses object.}

\item{mode}{String indicating the mode to use for the filtering. If 'keep' the proteins are selected and kept, while if 'remove' the proteins are discarded. Default: \code{"keep"}.}
}
\value{
An object of class \code{DEprot} (S4 vector).
}
\description{
Function that allows for removing or keeping specific proteins from any \code{DEprot} object class.
}
\examples{
# Remove a subset of proteins
dpo <- filter.proteins(DEprot.object = DEprot::test.toolbox$dpo.raw,
                       proteins = c("protein.29","protein.30"),
                       mode = "remove")

# Retain a subset of proteins
dpo <- filter.proteins(DEprot.object = DEprot::test.toolbox$dpo.raw,
                       proteins = c("protein.29","protein.30"),
                       mode = "keep")


}
\author{
Sebastian Gregoricchio
}
