% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.protein.info.R
\name{get.protein.info}
\alias{get.protein.info}
\title{get.protein.info}
\usage{
get.protein.info(DEprot.object)
}
\arguments{
\item{DEprot.object}{Any object of class \code{DEprot} or \code{DEprot.analyses}.}
}
\value{
Data.frame corresponding to the \code{protein.info} slot of the provided object (row names: protein IDs), or \code{NULL} when no annotation is stored.
}
\description{
Function to extract the protein annotation table from a \code{DEprot} object.
}
\examples{
# Extract the protein annotation from a DEprot object
get.protein.info(DEprot::dpo.imputed.counts)

# Extract the protein annotation from a DEprot.analyses object
get.protein.info(DEprot::test.toolbox$diff.exp.limma)

}
\seealso{
\link{add.protein.info}, \link{load.counts2}, \link{get.results}
}
\author{
Sebastian Gregoricchio
}
