% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_functions.R
\name{get.sPLSDA.proteins}
\alias{get.sPLSDA.proteins}
\title{get.sPLSDA.proteins}
\usage{
get.sPLSDA.proteins(
  DEprot.sPLSDA.object,
  component = NULL,
  mode = "union",
  direction = "both",
  top.n = NULL
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{component}{Numeric value (or vector) indicating the component(s) to use. Default: \code{NULL} (all the components).}

\item{mode}{String indicating how the components are combined. One among: 'union', 'intersect', 'list' (one vector per component). Default: \code{"union"}.}

\item{direction}{String indicating which side of the component to keep. One among: 'both', 'positive' (proteins higher in the reference group), 'negative'. Default: \code{"both"}.}

\item{top.n}{Numeric value indicating how many best-ranked proteins of each component should be returned. Default: \code{NULL} (all the selected proteins).}
}
\value{
A character vector, or a named list when \code{mode = "list"}.
}
\description{
Extracts the proteins retained by the sparse model. The output is a plain character vector, directly usable as \code{protein.subset} in \link{heatmap.counts} or as gene list of an over-representation analysis.
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

# All the selected proteins
get.sPLSDA.proteins(DEprot.sPLSDA.object = splsda)

# Top 3 proteins of the first component, higher in the reference group
get.sPLSDA.proteins(DEprot.sPLSDA.object = splsda,
                    component = 1,
                    direction = "positive",
                    top.n = 3)

}
\seealso{
\link{perform.sPLSDA}, \link{get.sPLSDA.results}, \link{sPLSDA.enrichment}
}
\author{
Sebastian Gregoricchio
}
