% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_functions.R
\name{get.sPLSDA.results}
\alias{get.sPLSDA.results}
\title{get.sPLSDA.results}
\usage{
get.sPLSDA.results(
  DEprot.sPLSDA.object,
  component = NULL,
  selected.only = TRUE,
  DEprot.object = NULL,
  protein.info.columns = "none",
  protein.info.prefix = NULL
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{component}{Numeric value (or vector) indicating the component(s) to return. Default: \code{NULL} (all the components).}

\item{selected.only}{Logical value indicating whether only the proteins retained by the sparse model should be returned. When \code{FALSE}, the complete loadings of the companion non-sparse model are returned instead. Default: \code{TRUE}.}

\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses} from which the protein annotation is taken. Default: \code{NULL} (no annotation appended).}

\item{protein.info.columns}{String vector indicating which columns of the \code{protein.info} slot should be appended. Two keywords are available: \code{"none"} and \code{"all"}. Default: \code{"none"}.}

\item{protein.info.prefix}{String to prepend to the names of the appended annotation columns. Default: \code{NULL}.}
}
\value{
A data.frame.
}
\description{
Simplifies the access to the loadings of a sPLS-DA. When the original object carries a protein annotation table (\code{protein.info} slot, see \link{load.counts2} and \link{add.protein.info}), the annotation columns can be appended.
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

# Proteins selected on the first component
get.sPLSDA.results(DEprot.sPLSDA.object = splsda, component = 1)

}
\seealso{
\link{perform.sPLSDA}, \link{get.sPLSDA.proteins}, \link{get.sPLSDA.ranking}
}
\author{
Sebastian Gregoricchio
}
