% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{get.timecourse.results}
\alias{get.timecourse.results}
\title{get.timecourse.results}
\usage{
get.timecourse.results(
  DEprot.timecourse.object,
  cluster = NULL,
  top.n = NULL,
  trending.only = TRUE
)
}
\arguments{
\item{DEprot.timecourse.object}{An object of class \code{DEprot.timecourse}.}

\item{cluster}{Numeric value (or vector) indicating the cluster(s) to retrieve. Default: \code{NULL} (all the clusters).}

\item{top.n}{Numeric value indicating how many best-ranked proteins should be kept (within the cluster when a \code{cluster} is provided, globally otherwise). Default: \code{NULL} (all the proteins).}

\item{trending.only}{Logic value indicating whether only the proteins showing a significant trend should be kept. Default: \code{TRUE}.}
}
\value{
A data.frame sorted by rank.
}
\description{
Retrieves the results table of a \code{DEprot.timecourse} object, optionally restricted to specific clusters and/or to the best-ranked proteins.
}
\author{
Sebastian Gregoricchio
}
