% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.contrasts.R
\name{heatmap.contrasts}
\alias{heatmap.contrasts}
\title{heatmap.contrasts}
\usage{
heatmap.contrasts(
  DEprot.analyses.object,
  contrasts = NULL,
  top.n = NULL,
  distance.method = "euclidean",
  clustering.method = "complete",
  high.color = "firebrick",
  low.color = "#2166AC",
  mid.color = "white",
  na.color = "gray",
  color.limits = c(NA, NA),
  cell.border.color = NA,
  cell.border.width = 0.5,
  title = NULL,
  show.protein.names = FALSE,
  protein.names.pattern = NULL,
  use.uncorrected.pvalue = FALSE
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrasts}{Numeric vector indicating the position of the contrast to use for the plotting. Default: \code{NULL} (all contrasts).}

\item{top.n}{Numeric value indicated the top differentially expressed proteins to consider for each contrast selected. The rank is based on the product of log2Fc and -log10Padj. Default: \code{NULL} (all differential proteins).}

\item{distance.method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given. Default: \code{"euclidean"}.}

\item{clustering.method}{The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Default: \code{complete"}.}

\item{high.color}{String indicating the color to use for up-regulated protein FoldChanges in the plot. Default: \code{"indianred"}.}

\item{low.color}{String indicating the color to use for down-regulated protein FoldChanges in the plot. Default: \code{"#2166AC"} (blue).}

\item{mid.color}{String indicating the color to use for the 0 in the plots. Default: \code{"white"}.}

\item{na.color}{String indicating the color to use for the NA values in the heatmap. Default: \code{"gray"}.}

\item{color.limits}{Numeric vector of length 2 indicating lower and upper limit of the color scale values. Default: \code{c(NA,NA)}, automatic limits applied.}

\item{cell.border.color}{String indicating the color to use for the individual cells border. Default: \code{NA} (no border).}

\item{cell.border.width}{Numeric value indicating the width of the cell border line. Ignored when \code{cell.border.color = NA}. Default: \code{0.5}.}

\item{title}{String indicating the title to use, markdown formatting supported. Default: \code{NULL} (automatic title).}

\item{show.protein.names}{Logical value to indicate whether the protein names should be displayed. Default: \code{FALSE}.}

\item{protein.names.pattern}{Character indicating a regular expression to remove from the protein IDs. Default" \code{NULL}, no alterations in the protein IDs.}

\item{use.uncorrected.pvalue}{Logical value indicating whether it should be used the normal p-value instead of the adjusted one (differential proteins numbers are recomputed). Default: \code{FALSE}, padj is used.}
}
\value{
A \code{DEprot.contrast.heatmap} object, which contains the heatmap (ggplot) and the hclust object used to order the rows.
}
\description{
Plots an heatmap of the log2(FoldChange) for differentially expressed proteins
}
