% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.distribution.R
\name{identify.distribution}
\alias{identify.distribution}
\title{identify.distribution}
\usage{
\method{identify}{distribution}(values)
}
\arguments{
\item{values}{Numeric vector of test statistic values (e.g., from a Wilcoxon or t-test). \code{NA} values are removed internally.}
}
\value{
A list with the following elements:
\describe{
  \item{\code{best}: }{String indicating the best-fitting distribution based on AIC.}
  \item{\code{comparison}: }{Data frame with AIC and BIC values for all successfully fitted distributions.}
  \item{\code{fits}: }{Named list of \code{fitdist} objects for each fitted distribution.}
  \item{\code{ks_test_best}: }{Kolmogorov-Smirnov test result for the best-fitting distribution.}
}
}
\description{
Identifies the best-fitting theoretical distribution (normal, t, chi-squared, or F)
for a given vector (such as test statistics) using maximum likelihood estimation.
Comparison is based on AIC (Akaike Information Criterion) and BIC (Bayesian Information Criterion) values,
and a Kolmogorov-Smirnov test is performed on the best-fitting distribution.
}
\examples{
data <- DEprot::test.toolbox$diff.exp.limma@analyses.result.list$condition_6h.10nM.E2.vs.6h.DMSO
result <- identify.distribution(values = data$results$statistic)

result$best
result$comparison

}
\author{
Sebastian Gregoricchio
}
