% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.counts.R
\name{impute.counts}
\alias{impute.counts}
\title{impute.counts}
\usage{
impute.counts(
  DEprot.object,
  method = "missForest",
  which.data = "randomized",
  overwrite.imputation = FALSE,
  missForest.max.iterations = 100,
  missForest.variable.wise.OOBerror = TRUE,
  missForest.cores = 1,
  missForest.parallel.mode = "variables",
  kNN.n.nearest.neighbours = 10,
  LLS.k = 2,
  pcaMethods.nPCs.to.test = 5,
  RegImpute.max.iterations = 10,
  RegImpute.fillmethod = "row_mean",
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{method}{String indicating the imputation method to use. One among: 'missForest', 'kNN' (VIM), 'tkNN' (imputomics), 'corkNN' (imputomics), 'LLS' (pcaMethods), 'SVD' (a.k.a svdImpute, pcaMethods), 'BPCA' (pcaMethods), 'PPCA' (pcaMethods), 'RegImpute' (DreamAI). Default: \code{"missForest"}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'randomized', 'random'. Default: \code{"randomized"}.}

\item{overwrite.imputation}{Logical value to indicate whether, in the case already available, the table of imputed counts should be overwritten. Default: \code{FALSE}.}

\item{missForest.max.iterations}{Max number of iterations for the missForest algorithm. Default: \code{100}.}

\item{missForest.variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{missForest.cores}{Number of cores used to run the \code{missForest} algorithm. If \code{missForest.cores} is higher than 1, the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{missForest.parallel.mode}. Default: \code{1}.}

\item{missForest.parallel.mode}{Define the mode to use for the parallelization, ignored when \code{missForest.cores} is 1 or lower. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{kNN.n.nearest.neighbours}{Numeric value indicating the number of nearest neighbors to use to perform the \code{kNN} imputation. Default: \code{10}.}

\item{LLS.k}{Cluster size, this is the number of similar genes used for regression. Default: \code{2}.}

\item{pcaMethods.nPCs.to.test}{Numeric value indicating the number of Principal Components to test in order to find the optimal number of PCs to used in the imputation methods from the \code{pcaMethods} package. This includes: 'LLS', 'SVD' (a.k.a 'svdImpute'), 'BPCA-pcaMethods', and 'PPCA'. Default: \code{5}.}

\item{RegImpute.max.iterations}{Numeric value indicating the number of maximum iteration for the imputation with \code{RegImpute} (from \code{DreamAI}). Default: \code{10}.}

\item{RegImpute.fillmethod}{String identifying the fill method to be used in the \code{RegImpute} method (from\code{DreamAI}). One among \code{"row_mean"} and \code{"zeros"}. Default: \code{"row_mean"}. It throws an warning if \code{"row_median"} is used.}

\item{seed}{Numeric value indicating the seed to use for the randomization. Default: \code{NULL}, automatically generated (saved in the \code{seed} element in the final imputation method list).}

\item{verbose}{Logical value indicating whether processing messages should be printed. Default: \code{FALSE}.}
}
\value{
A \code{DEprot} object. The boxplot showing the distribution of the protein intensity is remade and added to the slot (\code{boxplot.imputed}). A list with parameters and other info about the imputation is added as well in the \code{imputation} slot.
}
\description{
Function that allows for the imputation of missing values using 4 possible algorithms: \href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://rdrr.io/cran/VIM/man/kNN.html}{kNN}, \href{https://rdrr.io/bioc/pcaMethods/man/llsImpute.html}{LLS}, \href{https://rdrr.io/bioc/pcaMethods/man/pca.html}{SVD}.
}
\examples{
dpo <- impute.counts(DEprot.object = DEprot::test.toolbox$dpo.norm,
                     which.data = "normalized",
                     method = "bPCA")


}
\seealso{
\href{https://CRAN.R-project.org/package=missForest}{missForest}, \href{https://CRAN.R-project.org/package=VIM}{VIM}, \href{https://CRAN.R-project.org/package=pcaMethods}{pcaMethods} R-packages, \href{https://github.com/WangLab-MSSM/DreamAI}{DreamAI}, \href{https://github.com/BioGenies/imputomics}{imputomics}.
}
\author{
Sebastian Gregoricchio
}
