% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.propagation.R
\name{jaccardSim}
\alias{jaccardSim}
\title{jaccardSim}
\usage{
jaccardSim(idsInSet, score)
}
\value{
A list of similarity matrix \code{sim.mat} and input preference vector \code{ip.vec}.
}
\description{
Calculate Jaccard Similarity.
}
\author{
Zhiao Shi, Yuxing Liao
}
