% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.counts.R
\name{load.counts}
\alias{load.counts}
\title{load.counts}
\usage{
load.counts(
  counts,
  metadata,
  log.base,
  imputation = NA,
  normalization.method = NA,
  column.id = "column.id",
  protein.info = NULL,
  protein.info.id.column = NULL
)
}
\arguments{
\item{counts}{A data.frame or a matrix in which the rownames are the proteins and the columns the samples.}

\item{metadata}{A data.frame containing at least one column called \code{column.id} which corresponds to the colnames of \code{counts}. Any other column can be added and will correspond to a "feature"of each sample.}

\item{log.base}{Number indicating the base of the log used to transform the counts. If none transformation is applied, indicate the base \code{1}.}

\item{imputation}{A string indicating the imputation method used. If none, use the default value \code{NA}.}

\item{normalization.method}{String or list indicating the normalization method used. If none, use the default value \code{NA}.}

\item{column.id}{String indicating the name of the column to use as "column.id" from the metadata data.frame. This column must contain all the colnames of \code{counts}.}

\item{protein.info}{Optional \code{data.frame} (or \code{matrix}) with one row per protein and any number of annotation columns. The protein IDs must be provided as row names, in a column called \code{prot.id}, or in the column indicated by \code{protein.info.id.column}. Default: \code{NULL} (no annotation).}

\item{protein.info.id.column}{String indicating the name of the column of \code{protein.info} containing the protein IDs. If \code{NULL} (default), the row names are used.}
}
\value{
A \code{DEprot} object (S4 vector).
}
\description{
Function used to generate a \code{DEprot} object starting from counts and metadata. This function is deprecated use \link{load.counts2} instead.
}
\examples{
# Raw counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2)

# Normalized counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2,
                   normalization.method = "MaxQuant automatic")

# Imputed counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2,
                   imputation = "missForest")



}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
