% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.counts2.R
\name{load.counts2}
\alias{load.counts2}
\title{load.counts2}
\usage{
load.counts2(
  counts,
  metadata,
  data.type,
  log.base,
  normalization.method = NA,
  randomization.method = NA,
  imputation.method = NA,
  column.id = "column.id",
  protein.info = NULL,
  protein.info.id.column = NULL
)
}
\arguments{
\item{counts}{A data.frame or a matrix in which the rownames are the proteins and the columns the samples.}

\item{metadata}{A data.frame containing at least one column called \code{column.id} which corresponds to the colnames of \code{counts}. Any other column can be added and will correspond to a "feature"of each sample.}

\item{data.type}{String indicating the type of data that are loaded. One among: 'raw', 'normalized', 'randomized', 'imputed'.}

\item{log.base}{Number indicating the base of the log used to transform the counts. If none transformation is applied, indicate the base \code{1}. By default, the output data are always transformed in log2 base.}

\item{normalization.method}{String or list indicating the normalization method used. If none, use the default value \code{NA}.}

\item{randomization.method}{String or list indicating the randomization method used. If none, use the default value \code{NA}.}

\item{imputation.method}{A string indicating the imputation method used. If none, use the default value \code{NA}.}

\item{column.id}{String indicating the name of the column to use as "column.id" from the metadata data.frame. This column must contain all the colnames of \code{counts}.}

\item{protein.info}{Optional \code{data.frame} (or \code{matrix}) with one row per protein and any number of annotation columns (e.g., gene symbol, description, number of peptides). The protein IDs must be provided as row names, in a column called \code{prot.id}, or in the column indicated by \code{protein.info.id.column}; they must correspond to the row names of \code{counts}. The table is re-ordered to match the counts: proteins missing from the annotation are filled with \code{NA}, annotations of proteins absent from the counts are discarded. Default: \code{NULL} (no annotation).}

\item{protein.info.id.column}{String indicating the name of the column of \code{protein.info} containing the protein IDs. If \code{NULL} (default), the row names are used (or a column called \code{prot.id} when the table has no explicit row names).}
}
\value{
A \code{DEprot} object (S4 vector).
}
\description{
Function used to generate a \code{DEprot} object starting from counts and metadata.
}
\examples{
dpo <- load.counts2(counts = DEprot::unimputed.counts,
                    metadata = DEprot::sample.config,
                    log.base = 2,
                    data.type = "raw")


}
\seealso{
\link{add.protein.info}, \link{get.results}
}
\author{
Sebastian Gregoricchio
}
