% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NES.plot.R
\name{NES.plot}
\alias{NES.plot}
\title{NES.plot}
\usage{
NES.plot(
  enrichResult,
  pos.NES.label = NULL,
  neg.NES.label = NULL,
  pos.NES.color = "steelblue",
  neg.NES.color = "orange",
  string.pattern.to.remove = "HALLMARK|GOBP",
  alpha.range = c(0.3, 1),
  add.counts = TRUE,
  perc.bleeding.x = 8,
  axes.text.size = 10,
  title = "NES enrichments"
)
}
\arguments{
\item{enrichResult}{A \code{DEprot.enrichResult} object, as generated by \link{geneset.enrichment} DEprot function, or a \code{gseaResult} object as generate by \href{https://bioconductor.org/packages/release/bioc/html/clusterProfiler.html}{clusterprofiler}.}

\item{pos.NES.label}{String for the label of sets enriched in the group with positive ranking feature scores. Default: \code{NULL}, automatic value retrieved in the \code{DEprot.enrichResult} object otherwise \code{"+NES"}.}

\item{neg.NES.label}{String for the label of sets enriched in the group with negative ranking feature scores. Default: \code{NULL}, automatic value retrieved in the \code{DEprot.enrichResult} object otherwise \code{"-NES"}.}

\item{pos.NES.color}{String for the color of sets enriched in the group with positive ranking feature scores. Default: \code{"steelblue"}.}

\item{neg.NES.color}{String for the color of sets enriched in the group with negative ranking feature scores. Default: \code{"orange"}.}

\item{string.pattern.to.remove}{String with a regular expression of a pattern to be removed from the set names. Default: \code{"HALLMARK|GOBP"}.}

\item{alpha.range}{Numeric vector of length 2 indicating minimum and maximum value for the transparency. Individual values must be a number between 0 and 1. Default: \code{c(0.3, 1)}.}

\item{add.counts}{Logic value to indicate whether add labels with the set size counts. Default: \code{"TRUE"}.}

\item{perc.bleeding.x}{Numeric value indicating the percentage of the full x.axis range to add on the left on the right. Useful when labels are falling outside for the x-max. Default: \code{8} (\%).}

\item{axes.text.size}{Numeric value indicating the font size of the axis text. Default: \code{10}.}

\item{title}{String indicating the title of the plot. Default: \code{"NES enrichments"}.}
}
\value{
A ggplot object.
}
\description{
GSEA analyses NES plotter. Plotting of a bar plot indicating the Normalized Enrichment Score (NES) of a gsea result. Bars will have different colors by positive or negative enrichment and transparency proportional to the -log10 of their adjusted p-value, while the legend displays the adjusted p-values themselves.
}
\examples{
NES.plot(enrichResult = DEprot::test.toolbox$gsea.results)


}
\author{
Sebastian Gregoricchio
}
