% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.counts.R
\name{normalize.counts}
\alias{normalize.counts}
\title{normalize.counts}
\usage{
normalize.counts(
  DEprot.object,
  balancing.function = "median",
  NRI.RI.ratio.threshold = 0.5,
  overwrite.normalization = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{balancing.function}{A string indicating the function to use to balance the quantile normalization. Default: \code{"median"}.}

\item{NRI.RI.ratio.threshold}{Value indicating the NRI/RI (nearly rank invariant (NRI)/rank invariant (RI)) ratio threshold. Default: \code{0.5}.}

\item{overwrite.normalization}{Logical value to indicate whether, in the case already available, the table of normalized counts should be overwritten. Default: \code{FALSE}.}
}
\value{
A \code{DEprot} object. A boxplot showing the distribution of the protein intensity after normalization for each sample is added to the slot (\code{boxplot.norm}).
}
\description{
Function that allows for the tail-robust quantile normalization of the counts using the \href{https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/full/10.1002/pmic.202000068}{MBQN package}.
}
\seealso{
\link{load.counts}, \href{https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/full/10.1002/pmic.202000068}{MBQN package}.
}
\author{
Sebastian Gregoricchio
}
