% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot,ANY-method}
\alias{plot,DEprot,ANY-method}
\title{DEprot plot-method}
\usage{
\S4method{plot}{DEprot,ANY}(x, y, ..., ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of class \code{DEprot}.}

\item{y}{Not used.}

\item{...}{Further arguments passed to \code{patchwork::wrap_plots()} (e.g. \code{guides}, \code{widths}, \code{heights}, \code{design}). They must be named.}

\item{ncol, nrow}{The dimensions of the grid to create - if both are NULL (default) it will use the same logic as \code{facet_wrap()} to set the dimensions.}
}
\description{
DEprot plot-method
}
\keyword{internal}
