% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.normality,ANY-method}
\alias{plot,DEprot.normality,ANY-method}
\title{DEprot.normality plot-method}
\usage{
\S4method{plot}{DEprot.normality,ANY}(x, y, ..., n.samples = NULL)
}
\arguments{
\item{x}{Object of class \code{DEprot.normality}}

\item{y}{Not used.}

\item{...}{Further arguments passed to \code{patchwork::wrap_plots()} (e.g. \code{guides}, \code{widths}, \code{heights}, \code{design}). They must be named, and they are used only when several contrasts are assembled.}

\item{n.samples}{NUmber of samples to display (in order by metadata table). Default: \code{NULL} (all samples are shown).}
}
\description{
DEprot.normality plot-method
}
\keyword{internal}
