% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.outliers,ANY-method}
\alias{plot,DEprot.outliers,ANY-method}
\title{DEprot.outliers plot-method}
\usage{
\S4method{plot}{DEprot.outliers,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{DEprot.outliers}}

\item{y}{Not used.}

\item{...}{Not used.}
}
\description{
DEprot.outliers plot-method
}
\keyword{internal}
