% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.power,ANY-method}
\alias{plot,DEprot.power,ANY-method}
\title{DEprot.power plot-method}
\usage{
\S4method{plot}{DEprot.power,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{DEprot.power}.}

\item{y}{Not used.}

\item{...}{Passed to \code{patchwork::wrap_plots()}.}
}
\value{
The combination of all the plots available.
}
\description{
DEprot.power plot-method
}
