% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.sPLSDA,ANY-method}
\alias{plot,DEprot.sPLSDA,ANY-method}
\title{DEprot.sPLSDA plot-method}
\usage{
\S4method{plot}{DEprot.sPLSDA,ANY}(x, y, ..., what = "scatter")
}
\arguments{
\item{x}{Object of class \code{DEprot.sPLSDA}}

\item{y}{Not used.}

\item{...}{Further arguments passed to the corresponding plotting function.}

\item{what}{String indicating which plot to display. One among: 'scatter' (comp1-vs-comp2 and comp3-vs-comp2), 'cumulative', 'tuning', 'performance'. Default: \code{"scatter"}.}
}
\description{
DEprot.sPLSDA plot-method
}
\keyword{internal}
