% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.correlation.heatmap.R
\name{plot.correlation.heatmap}
\alias{plot.correlation.heatmap}
\title{plot.correlation.heatmap}
\usage{
\method{plot}{correlation.heatmap}(
  DEprot.object,
  correlation.method = "pearson",
  column.subset = NULL,
  use.normalized.data = TRUE,
  palette = viridis::mako(100, direction = -1),
  correlation.scale.limits = c(0, 1),
  exclude.diagonal = FALSE,
  dendrogram.position = "left",
  dendrogram.color = "black",
  dendrogram.linewidth = 0.5,
  display.values = TRUE,
  values.color = "white",
  values.decimals = 2,
  values.font.size = 2,
  values.transparency = 1,
  plot.title = paste0("**", stringr::str_to_title(correlation.method), " correlation**"),
  plot.subtitle = NULL,
  clustering.method = "complete"
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{correlation.method}{String indicating the clustering method to use to generate the correlation matrix. Possible options: 'pearson', 'spearman', 'kendall'. Default: \code{"pearson"}.}

\item{column.subset}{Vector indicating the name of the columns (\code{column.id} in the metadata table) to use/subset for the correlation. Default: \code{NULL} (no subsetting).}

\item{use.normalized.data}{Logical value to indicate whether to use the normalized data or not. Default: \code{TRUE}.}

\item{palette}{Vector of colors corresponding to the palette to use for the heatmap color scale. Default: \code{viridis::mako(100, direction = -1)}.}

\item{correlation.scale.limits}{Two-elements vector to indicate lower and higher limits, respectively, to apply to the correlation coefficient color scale. Default: \code{c(0,1)}.}

\item{dendrogram.position}{String indicating the position of the dendrogram. One among: "top", "bottom", "left", "right". Default: \code{"left"}.}

\item{dendrogram.color}{String indicating the color of the dendrogram lines. Default: \code{"black"}.}

\item{dendrogram.linewidth}{Numeric value indicating the line.width of the dendrogram. Default: \code{"black"}.}

\item{display.values}{Logical value indicating whether the correlation coefficient should be displayed for each cell. Default: \code{TRUE}.}

\item{values.color}{String indicating the color to use for the correlation coefficient labels. Default: \code{"white"}.}

\item{values.decimals}{Numeric value indicating the number of decimals at which round the correlation coefficient labels. Default: \code{2}.}

\item{values.font.size}{Numeric value indicating the font size of the correlation coefficient labels. Default: \code{2}.}

\item{values.transparency}{Numeric value between 0-1 indicating the transparency (alpha) of the correlation coefficient labels. Default: \code{1}, full color.}

\item{plot.title}{String indicating the main title of the plot. Default: \code{paste(stringr::str_to_title(correlation.method), "correlation")}.}

\item{plot.subtitle}{String indicating the subtitle of the plot. Default: \code{NULL}.}

\item{clustering.method}{String indicating the clustering method to use. The value should be (an unambiguous abbreviation of) one among: 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA), 'mcquitty' (= WPGMA), 'median' (= WPGMC) or 'centroid' (= UPGMC).}
}
\value{
A \code{DEprot.correlation} with the correlation heatmap in ggplot format.
}
\description{
Function to generate a correlation heatmap. Includes dendrogram and clustering of the data.
}
