% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.counts.R
\name{plot.counts}
\alias{plot.counts}
\title{plot.counts}
\usage{
\method{plot}{counts}(
  DEprot.object,
  which.data = "imputed",
  violin.color = "darkorange",
  max.line.color = "indianred",
  min.line.color = "steelblue",
  title = NULL,
  subtitle = NULL,
  convert.log2 = TRUE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot.analyses}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{violin.color}{String indicating the color to use for the violin plots. Default: \code{"darkorange"}.}

\item{max.line.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"indianred"}.}

\item{min.line.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"steelblue"}.}

\item{title}{String indicating the title to use. Default: \code{NULL}.}

\item{subtitle}{String indicating the subtitle to use. Default: \code{NULL}.}

\item{convert.log2}{Logical value to define whether counts should be log2 transformed. Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plots a volcano plot log2(FoldChange) x -log10(p.adjusted) of differential expression results
}
\examples{
plot.counts(DEprot.object = DEprot::test.toolbox$dpo.imp)


}
\author{
Sebastian Gregoricchio
}
