% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GSEA.R
\name{plot.GSEA}
\alias{plot.GSEA}
\title{plot.GSEA}
\usage{
\method{plot}{GSEA}(
  gsea.results,
  geneset.id = NULL,
  enrichment.geom = "line",
  enrichment.color = "green",
  enrichment.geom.size = 1,
  enrichment.plot.zero.line = TRUE,
  enrichment.zero.line.color = "gray",
  enrichment.zero.line.width = 0.5,
  enrichment.annotations.vjust.offset = 0,
  geneset.segments.width = 0.3,
  geneset.segments.color = "black",
  add.condition.labels = TRUE,
  condition.label.positive = NULL,
  condition.label.negative = NULL,
  rank.max.color = "indianred",
  ranking.color = "gray",
  gradient.colors = c("Reds", "Blues"),
  title.position = "center",
  title = NA,
  image.file.name = NULL,
  image.width = 7,
  image.height = 5,
  return.all.objects = FALSE
)
}
\arguments{
\item{gsea.results}{Object of class \code{DEprot.enrichResult}, generated using \link{geneset.enrichment}, or \code{gseaResult}, generate by \href{https://bioconductor.org/packages/release/bioc/html/clusterProfiler.html}{clusterprofiler}.}

\item{geneset.id}{Numeric value or a string identifying the Nth geneSet (numeric) or a specific id (string) if geneSet in the result table. Default value: \code{NULL}, which returns the ordered list of available geneSets.}

\item{enrichment.geom}{String indicating the type of graph to use to plot the enrichment scores. Possible options: 'line', 'lines', 'dot', 'dots', 'point', 'points' (case insensitive). Default: \code{'line'}.}

\item{enrichment.color}{String indicating any R-supported color to be used for the enrichment score plot. Default: \code{'green'}.}

\item{enrichment.geom.size}{Numeric value indicating the size of the line, or dots, used in the enrichment score plot. Default: \code{1}.}

\item{enrichment.plot.zero.line}{Logical value to indicated whether to plot an horizontal line at 0 in the enrichment score plot. Default: \code{TRUE}.}

\item{enrichment.zero.line.color}{String indicating any R-supported color to be used for the 0-line in the enrichment score plot (active when \code{enrichment.plot.zero.line = TRUE}). Default: \code{'gray'}.}

\item{enrichment.zero.line.width}{Numeric value indicating the line width of the 0-line in the enrichment score plot (active when \code{enrichment.plot.zero.line = TRUE}). Default: \code{0.5}.}

\item{enrichment.annotations.vjust.offset}{Numeric value to add to the vjust (vertical positioning) of the enrichment plot annotations (P, Padj, q, NES, set size). Positive values will shift-down the annotations. Default: \code{0} (base line).}

\item{geneset.segments.width}{Numeric value indicating the line width of the geneSet vertical segments. Default: \code{0.3}.}

\item{geneset.segments.color}{String indicating any R-supported color to be used for the geneSet segments. Default: \code{'black'}.}

\item{add.condition.labels}{Logic value indicating whether the labels of the conditions compared should be plotted. Ignored if the object is not of class \code{DEprot.enrichResult}. Default: \code{TRUE}.}

\item{condition.label.positive}{String indicating the label to use for the condition showing a positive ranking score (left side of the plot). Ignored if \code{add.condition.labels = FALSE}. Default: \code{NULL}, none or automatic.}

\item{condition.label.negative}{String indicating the label to use for the condition showing a negative ranking score (right side of the plot). Ignored if \code{add.condition.labels = FALSE}. Default: \code{NULL}, none or automatic.}

\item{rank.max.color}{String indicating any R-supported color to be used for the max rank dotted lines and annotation. Default: \code{'indianred'}.}

\item{ranking.color}{String indicating any R-supported color to be used for the ranked list plot (histogram). Default: \code{'gray'}.}

\item{gradient.colors}{Two-values string vector indicating the shadows of palettes to use for the genset gradient. Possible values: 'Blues', 'Greens', 'Greys', 'Oranges', 'Purples', 'Reds'. Default: \code{c('Reds', 'Blues')}.}

\item{title.position}{String indicating the position of the title: 'left', 'center', 'right'. Default: \code{'center'}.}

\item{title}{String indicating the title to use. Default: \code{NA}, this will automatically use the geneset name chosen. Use \code{NULL} to do not plot the title.}

\item{image.file.name}{String indicating the full path for the export of a pdf file of the combined plot. Default: \code{NULL}, no plot will be exported.}

\item{image.width}{Numeric value to indicate the width (in inches) to use for the exported pdf file. Active only when \code{image.file.name} is not \code{NULL}. Default: \code{7}.}

\item{image.height}{Numeric value to indicate the height (in inches) t use for the exported pdf file. Active only when \code{image.file.name} is not \code{NULL}. Default: \code{5}.}

\item{return.all.objects}{Logical value to indicate whether the function should return only the combined plot (ggplot object), or all the different panels and the combined plot in a list. Default: \code{FALSE} (only combined plot).}
}
\value{
Either a ggplot-object with the final combined plot, or a list with the three panels separated and the combined plot: \code{list(enrichment.panel, geneset.panel, rank.panel, combined.plot)}.
}
\description{
Function to plot GSEA results (see \href{https://bioconductor.org/packages/release/bioc/html/clusterProfiler.html}{clusterprofiler}).
}
\seealso{
The internal function \code{gsInfo} is derived from the package \href{https://github.com/YuLab-SMU/enrichplot/blob/bf86dcdc74316c7ee8d9e7f5fcac7e88392204e6/R/gseaplot.R#L86}{enrichplot::gsInfo}.
}
\author{
Sebastian Gregoricchio
}
