% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PCoA.scatter.R
\name{plot.PCoA.scatter}
\alias{plot.PCoA.scatter}
\title{plot.PCoA.scatter}
\usage{
\method{plot}{PCoA.scatter}(
  DEprot.PCoA.object,
  PCo.x = 2,
  PCo.y = 1,
  color.column = "column.id",
  shape.column = NULL,
  label.column = NULL,
  plot.zero.line.x = TRUE,
  plot.zero.line.y = TRUE
)
}
\arguments{
\item{DEprot.PCoA.object}{An object of class \code{DEprot.PCoA}, as generated by \link{perform.PCoA}.}

\item{PCo.x}{Number indicating which Principal Coordinate (PCo) to display on the x-axis. Default: \code{2} (PCo2).}

\item{PCo.y}{Number indicating which Principal Coordinate (PCo) to display on the y-axis. Default: \code{1} (PCo1).}

\item{color.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot colors. Default: \code{"column.id"} (each sample a color).}

\item{shape.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the \code{metadata} to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{plot.zero.line.x}{Logical value to indicate whether to plot a gray dashed line in correspondence of x=0. Default: \code{TRUE}.}

\item{plot.zero.line.y}{Logical value to indicate whether to plot a gray dashed line in correspondence of y=0. Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plot a scatter of the principal coordinates from an object generated by \link{perform.PCoA}.
}
\details{
The axis labels report the proportion of variance of each principal coordinate, computed using only the
positive eigenvalues as denominator (see \link{perform.PCoA}). Axes corresponding to a negative eigenvalue
have no real representation and raise an error.
}
\examples{
# Compute the sample correlations and the PCoA
correlation <- plot.correlation.heatmap(DEprot.object = DEprot::test.toolbox$dpo.imp,
                                        correlation.method = "spearman")

pcoa <- perform.PCoA(DEprot.correlation.object = correlation)

# Plot the PCo1-vs-PCo2 scatter
plot.PCoA.scatter(DEprot.PCoA.object = pcoa,
                  PCo.x = 1,
                  PCo.y = 2,
                  color.column = "condition",
                  shape.column = "replicate",
                  label.column = "replicate")

}
\seealso{
\link{perform.PCoA}, \link{plot.PCoA.scatter.123}, \link{plot.PC.scatter}
}
\author{
Sebastian Gregoricchio
}
