% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.biplot}
\alias{plot.sPLSDA.biplot}
\title{plot.sPLSDA.biplot}
\usage{
\method{plot}{sPLSDA.biplot}(
  DEprot.sPLSDA.object,
  comp.x = 1,
  comp.y = 2,
  color.column = NULL,
  shape.column = NULL,
  label.column = NULL,
  ellipse = FALSE,
  ellipse.level = 0.95,
  plot.zero.line.x = TRUE,
  plot.zero.line.y = TRUE,
  n.loadings = 10,
  loading.color = "turquoise4",
  loading.arrow.size = 0.6,
  loading.label.size = 3,
  loading.alpha = 0.7,
  loading.scale = 0.8
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{comp.x}{Number indicating which component to display on the x-axis. Default: \code{1}.}

\item{comp.y}{Number indicating which component to display on the y-axis. Default: \code{2}.}

\item{color.column}{String indicating the name of the column in the metadata to use as factor for the dot colors. Default: \code{NULL} (the column used to fit the model).}

\item{shape.column}{String indicating the name of the column in the metadata to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the metadata to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{ellipse}{Logical value indicating whether a confidence ellipse should be drawn around each group. Default: \code{FALSE}.}

\item{ellipse.level}{Numeric value (0-1) indicating the confidence level of the ellipses. Default: \code{0.95}.}

\item{plot.zero.line.x}{Logical value to indicate whether to plot a gray dashed line in correspondence of x=0. Default: \code{TRUE}.}

\item{plot.zero.line.y}{Logical value to indicate whether to plot a gray dashed line in correspondence of y=0. Default: \code{TRUE}.}

\item{n.loadings}{Number of proteins to display, ranked by their distance from the origin in the plane displayed. Default: \code{10}.}

\item{loading.color}{String indicating the color of the arrows and of their labels. Default: \code{"turquoise4"}.}

\item{loading.arrow.size}{Numeric value indicating the linewidth of the arrows. Default: \code{0.6}.}

\item{loading.label.size}{Numeric value indicating the font size of the labels. Default: \code{3}.}

\item{loading.alpha}{Numeric value (0-1) indicating the transparency of arrows and labels. Default: \code{0.7}.}

\item{loading.scale}{Numeric multiplier to manually adjust the length of the arrows. Default: \code{0.8}.}
}
\value{
A ggplot object.
}
\description{
Overlays the loadings of the selected proteins on the scatter of two components of a sPLS-DA.
}
\details{
Only the proteins retained on at least one of the two components displayed can appear: a protein dropped by the sparsity penalty has a coefficient of exactly zero and no direction to point to.
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

plot.sPLSDA.biplot(DEprot.sPLSDA.object = splsda,
                   comp.x = 1,
                   comp.y = 2,
                   n.loadings = 5)

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.scatter}, \link{plot.sPLSDA.loadings}
}
\author{
Sebastian Gregoricchio
}
