% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.loadings}
\alias{plot.sPLSDA.loadings}
\title{plot.sPLSDA.loadings}
\usage{
\method{plot}{sPLSDA.loadings}(
  DEprot.sPLSDA.object,
  component = 1,
  n.proteins = 25,
  group.colors = NULL,
  protein.names.pattern = NULL,
  bar.width = 0.75,
  axes.text.size = 9,
  title = NULL
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{component}{Numeric value indicating the component to display. Default: \code{1}.}

\item{n.proteins}{Numeric value indicating the maximum number of proteins to display, ranked by absolute loading. Default: \code{25}.}

\item{group.colors}{Color-vector (optionally named on the classes) used for the bars. Default: \code{NULL} (automatic colors).}

\item{protein.names.pattern}{String indicating a pattern to be passed to gsub and removed from the protein names. Default: \code{NULL} (no changes).}

\item{bar.width}{Numeric value indicating the width of the bars. Default: \code{0.75}.}

\item{axes.text.size}{Numeric value indicating the font size of the axis text. Default: \code{9}.}

\item{title}{String indicating the title of the plot (markdown annotation supported). Default: \code{NULL} (automatic title).}
}
\value{
A ggplot object.
}
\description{
Plots the loadings of the proteins retained on a component of a sPLS-DA, colored by the class in which each of them is the highest.
}
\details{
The sign of a loading is only interpretable once the orientation of the component is fixed, which \link{perform.sPLSDA} does through the \code{reference.group} argument: a positive coefficient marks a protein pulling the samples towards that group. The color adds the information the coefficient alone does not carry, namely in which class the protein is actually the most abundant, assigned by the highest median as done by \code{mixOmics::plotLoadings}.
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

plot.sPLSDA.loadings(DEprot.sPLSDA.object = splsda,
                     component = 1,
                     n.proteins = 5)

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.stability}, \link{get.sPLSDA.results}
}
\author{
Sebastian Gregoricchio
}
