% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.upset.R
\name{plot.upset}
\alias{plot.upset}
\title{plot.upset}
\usage{
\method{plot}{upset}(
  DEprot.analyses.object,
  contrast.subset = NULL,
  title = NULL,
  subtitle = NULL,
  sort.intersections = "cardinality",
  sort.sets = "descending",
  intersection.bar.color = "black",
  setsize.bar.color = "black",
  show.counts = TRUE,
  height.ratio = 0.5,
  width.ratio = 0.3,
  use.uncorrected.pvalue = FALSE
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast.subset}{Numeric vector indicating the contrasts to use. Default: \code{NULL} (all contrasts are shown).}

\item{title}{String indicating the title to display. Markdown-mode supported. Default: \code{NULL} (no title).}

\item{subtitle}{String indicating the title to display. Markdown-mode supported. Default: \code{NULL} (subtitle).}

\item{sort.intersections}{String indicating which method to use for the intersection sorting: 'cardinality' or 'degree'. Default: \code{"cardinality"}.}

\item{sort.sets}{String indicating which method to use for the sets sorting: 'descending' or 'ascending'. Default: \code{"descending"}.}

\item{intersection.bar.color}{String indicating the color to use for the intersection size bar plot. Default: \code{"black"}.}

\item{setsize.bar.color}{String indicating the color to use for the set size bar plot. Default: \code{"black"}.}

\item{show.counts}{Logical value indicating whether the counts in the intersection bar pot should be shown. Default: \code{TRUE}.}

\item{height.ratio}{Numeric value indicating the ratio of the intersection matrix to intersection size height. Default: \code{0.5}.}

\item{width.ratio}{Numeric value indicating the ratio of the overall set size width to intersection matrix width. Default: \code{0.3}.}

\item{use.uncorrected.pvalue}{Logical value indicating whether it should be used the normal p-value instead of the adjusted one (differential proteins numbers are recomputed). Default: \code{FALSE}, padj is used.}
}
\value{
A \code{DEprot.upset} with the upset plot and the matrix of the observed enriched proteins.
}
\description{
Plots an upset-plot for differential expression results
}
