% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.volcano.R
\name{plot.volcano}
\alias{plot.volcano}
\title{plot.volcano}
\usage{
\method{plot}{volcano}(
  DEprot.analyses.object,
  contrast = 1,
  up.color = "indianred",
  down.color = "steelblue",
  unresponsive.color = "purple",
  null.color = "gray",
  point.size = 2,
  point.alpha = 0.5,
  title = NULL,
  use.uncorrected.pvalue = FALSE,
  symmetric.x = TRUE,
  dot.labels = NULL,
  protein.names.pattern = "",
  labels.in.boxes = FALSE,
  label.font.size = 2,
  label.max.overlaps = 100,
  min.segment.length.labels = 0,
  label.top.n = NULL
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Number indicating the position of the contrast to use for the plotting.}

\item{up.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"indianred"}.}

\item{down.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"steelblue"}.}

\item{unresponsive.color}{String indicating the color to use for unresponsive proteins in the plots. Default: \code{"purple"}.}

\item{null.color}{String indicating the color to use for null proteins in the plots. Default: \code{"gray"}.}

\item{point.size}{Numeric value indicating the size of the dots. Default: \code{2}.}

\item{point.alpha}{Numeric value between 0 and 1 to indicate the transparency (alpha) of the dots. Default: \code{0.5}.}

\item{title}{String indicating the title to use. Default: \code{NULL} (automatic title).}

\item{use.uncorrected.pvalue}{Logical value indicating whether it should be used the normal p-value instead of the adjusted one (differential proteins numbers are recomputed). Default: \code{FALSE}, padj is used.}

\item{symmetric.x}{Logical values indicating whether the x-axis scale should be symmetric or not. Default: \code{TRUE}.}

\item{dot.labels}{String vector indicating labels to show on the plot that should correspond to \code{prot.id} column values. Default: \code{NULL} (no labels shown).}

\item{protein.names.pattern}{Character indicating a regular expression to remove from the protein IDs. Default: \code{""}, no alterations in the protein IDs.}

\item{labels.in.boxes}{Logical value indicating whether the labels should be visualized as boxes. Default: \code{FALSE}.}

\item{label.font.size}{Numeric value indicating the size to use for the dot labels. Default: \code{2}.}

\item{label.max.overlaps}{Numeric value indicating the maximum number of overlaps allowed between labels. Default: \code{100}.}

\item{min.segment.length.labels}{Numeric value indicating the minimal length of the segments that connect the labels to the points. Default: \code{0} (segment always shown).}

\item{label.top.n}{Single integer value (or \code{NULL}) indicating the number of top differentially expressed proteins to label automatically. Differential proteins (up- or down-regulated) are ranked by \code{-log10(padj) * abs(log2FC)} and the top \code{N} are labeled; if fewer than \code{N} differential proteins are available, all of them are labeled. When \code{use.uncorrected.pvalue = TRUE}, the uncorrected p-value is used in the ranking instead of the adjusted one. Any IDs provided through \code{dot.labels} are added to the automatically selected ones. The \code{protein.names.pattern}, \code{labels.in.boxes} and other label aesthetics are applied to these labels as well. Default: \code{NULL} (no automatic labels).}
}
\value{
A ggplot object.
}
\description{
Plots a volcano plot log2(FoldChange) x -log10(p.adjusted) of differential expression results
}
\examples{
# With specific-labels
plot.volcano(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
             contrast = 1,
             dot.labels = "protein.19",
             labels.in.boxes = FALSE) +
  ggplot2::xlab("log<sub>2</sub>(Fold Change FBS/6h.DMSO)")


# Labeling top-10 proteins
plot.volcano(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
             contrast = 1,
             dot.labels = "protein.19",
             label.top.n = 10,
             labels.in.boxes = FALSE) +
  ggplot2::xlab("log<sub>2</sub>(Fold Change FBS/6h.DMSO)")

}
\author{
Sebastian Gregoricchio
}
