% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.missing.values.R
\name{randomize.missing.values}
\alias{randomize.missing.values}
\title{randomize.missing.values}
\usage{
randomize.missing.values(DEprot.object, group.column, tail.percentage = 5)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{group.column}{String indicating a column for the metadata table which indicates the IDs of the conditions within which the missing values should be tested.}

\item{tail.percentage}{Numeric value between 0 and 100 (excluded) indicating the threshold percentage of the distribution from which random values should be sampled. Default: \code{5} (percent).}
}
\value{
A \code{DEprot} object in which the completely missing values have been substituted in the normalized counts table.
}
\description{
Function that allows to assign random values (from the bottom-X percent of the LFQ values distribution) to proteins that show missing values for all the replicates of a specific condition.
}
