% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.undetected.proteins.R
\name{remove.undetected.proteins}
\alias{remove.undetected.proteins}
\title{remove.undetected.proteins}
\usage{
remove.undetected.proteins(
  DEprot.object,
  min.n.samples = 3,
  which.data = "normalized"
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot} object.}

\item{min.n.samples}{Numeric value indicating the minimum number of sample in which a protein must be detected to be kept. Default: \code{3}.}

\item{which.data}{String indicating which type of counts should be used to define the proteins to remove. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"normalized"}.}
}
\value{
A \code{DEprot} object.
}
\description{
This function allows for the removal of all the proteins that occur below a certain threshold (detected in at least n samples).
}
\examples{
dpo <- load.counts(counts = DEprot::unimputed.counts, metadata = DEprot::sample.config, log.base = 2)
dpo <- remove.undetected.proteins(DEprot.object = dpo, min.n.samples = 3, which.data = "raw")


}
\author{
Sebastian Gregoricchio
}
