% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.bait.R
\name{rescale.bait}
\alias{rescale.bait}
\title{rescale.bait}
\usage{
rescale.bait(DEprot.object, bait.id)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses} (must contain imputed counts).}

\item{bait.id}{String indicating the id of a protein of the dataset present in the DEprot object.}
}
\value{
An object of class \code{DEprot} or \code{DEprot.analyses} (same class of the input).
}
\description{
Rescaling of the imputed counts based on the mean of a specific protein. This is useful for comparing the enrichment of interactors in the RIME experiments.
}
\examples{
recaled_dpo <- rescale.bait(DEprot.object = DEprot::test.toolbox$dpo.imp, bait.id = "protein.29")

}
\author{
Sebastian Gregoricchio
}
