% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rime.dpo}
\alias{rime.dpo}
\title{Imputed RIME (LFQ) data}
\format{
Object of class \code{DEprot}
}
\source{
Eickhoff et al. (Commun Biol, 2025)
}
\usage{
data("rime.dpo")
}
\description{
Subset of data from Eickhoff et al. (Commun Biol, 2025) that have been reanalyzed in Eickhoff et al. (NAR Genom Bioinform, 2026).
}
\author{
Nils Eickhoff and Sebastian Gregoricchio
}
\keyword{datasets}
