% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rime.saintq}
\alias{rime.saintq}
\title{SAINTq results for RIME}
\format{
Object of class \code{DEprot.SAINTq}
}
\source{
Eickhoff et al. (Commun Biol, 2025)
}
\usage{
data("rime.saintq")
}
\description{
Results of SAINTq (original tool) run on a subset of data from Eickhoff et al. (Commun Biol, 2025) that have been reanalyzed in Eickhoff et al. (NAR Genom Bioinform, 2026).
}
\author{
Nils Eickhoff and Sebastian Gregoricchio
}
\keyword{datasets}
