% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.config}
\alias{sample.config}
\title{Example of metadata table}
\format{
A data frame with 24 rows and 7 columns:
\describe{
  \item{\code{rows}}{proteins}
  \item{\code{columns}}{samples}
}
}
\source{
Simulated data
}
\usage{
sample.config
}
\description{
Dummy table for metadata corresponding to the sample configuratiuon of \link{unimputed.counts}.
}
\keyword{datasets}
