% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot-method}
\alias{show,DEprot-method}
\title{DEprot show-method}
\usage{
\S4method{show}{DEprot}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot}}
}
\description{
DEprot show-method
}
