% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.analyses-method}
\alias{show,DEprot.analyses-method}
\title{DEprot.analyses show-method}
\usage{
\S4method{show}{DEprot.analyses}(object)
}
\description{
DEprot.analyses show-method
}
