% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.contrast.heatmap-method}
\alias{show,DEprot.contrast.heatmap-method}
\title{DEprot.cluster.heatmap show-method}
\usage{
\S4method{show}{DEprot.contrast.heatmap}(object)
}
\description{
DEprot.cluster.heatmap show-method
}
