% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.correlation-method}
\alias{show,DEprot.correlation-method}
\title{DEprot.correlation show-method}
\usage{
\S4method{show}{DEprot.correlation}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.correlation}}
}
\description{
DEprot.correlation show-method
}
