% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.counts.heatmap-method}
\alias{show,DEprot.counts.heatmap-method}
\title{DEprot.counts.heatmap show-method}
\usage{
\S4method{show}{DEprot.counts.heatmap}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.counts.heatmap}}
}
\description{
DEprot.counts.heatmap show-method
}
