% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.enrichResult-method}
\alias{show,DEprot.enrichResult-method}
\title{DEprot.enrichResult show-method}
\usage{
\S4method{show}{DEprot.enrichResult}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.enrichResult}}
}
\description{
DEprot.enrichResult show-method
}
