% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.missingness-method}
\alias{show,DEprot.missingness-method}
\title{DEprot.missingness show-method}
\usage{
\S4method{show}{DEprot.missingness}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.missingness}}
}
\description{
DEprot.missingness show-method
}
