% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.normality-method}
\alias{show,DEprot.normality-method}
\title{DEprot.normality show-method}
\usage{
\S4method{show}{DEprot.normality}(object)
}
\description{
DEprot.normality show-method
}
