% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.outliers-method}
\alias{show,DEprot.outliers-method}
\title{DEprot.outliers show-method}
\usage{
\S4method{show}{DEprot.outliers}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.outliers}}
}
\description{
DEprot.outliers show-method
}
