% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.PCoA-method}
\alias{show,DEprot.PCoA-method}
\title{DEprot.PCoA show-method}
\usage{
\S4method{show}{DEprot.PCoA}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.PCoA}}
}
\description{
DEprot.PCoA show-method
}
