% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.power-method}
\alias{show,DEprot.power-method}
\title{DEprot.power show-method}
\usage{
\S4method{show}{DEprot.power}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.power}.}
}
\description{
DEprot.power show-method
}
