% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.RMSE-method}
\alias{show,DEprot.RMSE-method}
\title{DEprot.RMSE show-method}
\usage{
\S4method{show}{DEprot.RMSE}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.RMSE}}
}
\description{
DEprot.RMSE show-method
}
