% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.SAINTq-method}
\alias{show,DEprot.SAINTq-method}
\title{DEprot.SAINTq show-method}
\usage{
\S4method{show}{DEprot.SAINTq}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.SAINTq}}
}
\description{
DEprot.SAINTq show-method
}
