% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.sPLSDA-method}
\alias{show,DEprot.sPLSDA-method}
\title{DEprot.sPLSDA show-method}
\usage{
\S4method{show}{DEprot.sPLSDA}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.sPLSDA}}
}
\description{
DEprot.sPLSDA show-method
}
