% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.enrichments.R
\name{simplify.enrichment}
\alias{simplify.enrichment}
\title{simplify.enrichment}
\usage{
simplify.enrichment(enrichment.results, dpo.analyses.object)
}
\arguments{
\item{enrichment.results}{Object of class \code{DEprot.enrichResult} as generated using \link{geneset.enrichment}.}

\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}
}
\value{
An object of class \code{DEprot.enrichResult}.
}
\description{
Perform Gene Set Enrichment Analyses (GSEA) or OverRepresentation Analyses (ORA) on a proteomics differential analyses.
}
