% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.missingness-method}
\alias{summary,DEprot.missingness-method}
\title{DEprot.missingness summary-method}
\usage{
\S4method{summary}{DEprot.missingness}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.missingness}}
}
\description{
DEprot.missingness summary-method
}
