% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.RMSE-method}
\alias{summary,DEprot.RMSE-method}
\title{DEprot.RMSE summary-method}
\usage{
\S4method{summary}{DEprot.RMSE}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.RMSE}}
}
\description{
DEprot.RMSE summary-method
}
