% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.SAINTq-method}
\alias{summary,DEprot.SAINTq-method}
\title{DEprot.SAINTq summary-method}
\usage{
\S4method{summary}{DEprot.SAINTq}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.SAINTq}}
}
\description{
DEprot.SAINTq summary-method
}
