% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.sPLSDA-method}
\alias{summary,DEprot.sPLSDA-method}
\title{DEprot.sPLSDA summary-method}
\usage{
\S4method{summary}{DEprot.sPLSDA}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.sPLSDA}}
}
\description{
DEprot.sPLSDA summary-method
}
