% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.toolbox}
\alias{test.toolbox}
\title{Test toolbox}
\format{
A list of objects:
\describe{
  \item{\code{test.unimputed.lfq}}{table of a subset of 50 proteins from the \link{unimputed.counts} table. Usable in combination with \link{sample.config}}
  \item{\code{dpo.raw}}{Object of class \code{DEprot} with unimputed data loaded as raw}
  \item{\code{dpo.norm}}{Object of class \code{DEprot} containing MBQN normalized data}
  \item{\code{dpo.imp}}{Object of class \code{DEprot} containing missForest imputed data}
  \item{\code{diff.exp.limma}}{Object of class \code{DEprot.analyses} containing \code{limma} defined differential expression for two contrasts}
  \item{\code{geneset}}{Dummy geneset TERM2GENE table for GSEA and ORA analyses containing 3 sets of proteins}
  \item{\code{gsea.results}}{Object of class \code{DEprot.enrichResult} result of GSEA analyses on contrast 1}
  \item{\code{ora.results}}{Object of class \code{DEprot.enrichResult} result of ORA analyses on contrast 1 and differential category '6h.DMSO'}
}
}
\source{
Simulated data
}
\usage{
data("test.toolbox")
}
\description{
Set of objects and datasets used for the package tests.
}
\author{
Sebastian Gregoricchio
}
\keyword{datasets}
