% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unimputed.counts}
\alias{unimputed.counts}
\title{Unimputed proteomics (LFQ) data}
\format{
A data frame with 13,239 rows and 12 columns:
\describe{
  \item{\code{rows}}{proteins}
  \item{\code{columns}}{samples}
}

Object of class \code{DEprot}
}
\source{
Simulated data

Simulated data
}
\usage{
unimputed.counts

unimputed.counts
}
\description{
Dummy example of full proteomics data (LFQ values in log2). Not imputed.

Dummy example of full proteomics data (LFQ values in log2). Imputed using missForest algorithm.
}
\keyword{datasets}
