% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.imp.methods.R
\name{compare.imp.methods}
\alias{compare.imp.methods}
\title{compare.imp.methods}
\usage{
\method{compare}{imp.methods}(
  DEprot.object,
  percentage.test = 30,
  correlation.method = "pearson",
  sample.group.column = NULL,
  which.data = "randomized",
  masking = "MCAR",
  max.NA.per.row = 0,
  max.NA.fraction = 0.5,
  dropout.by = "group",
  run.missForest = TRUE,
  run.kNN = TRUE,
  run.tkNN = TRUE,
  run.corkNN = TRUE,
  run.LLS = TRUE,
  run.SVD = TRUE,
  run.BPCA = TRUE,
  run.PPCA = TRUE,
  run.RegImpute = TRUE,
  missForest.max.iterations = 100,
  missForest.variable.wise.OOBerror = TRUE,
  missForest.cores = 1,
  missForest.parallel.mode = "variables",
  kNN.n.nearest.neighbours = 10,
  LLS.k = 2,
  pcaMethods.nPCs.to.test = 5,
  RegImpute.max.iterations = 10,
  RegImpute.fillmethod = "row_mean",
  normalize.color.bar = TRUE,
  low.residual.color = "firebrick",
  zero.residual.color = "white",
  high.residual.color = "steelblue4",
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts} or \link{load.counts2}.}

\item{percentage.test}{Numeric value between 0 (excluded) and 100 indicating the percentage of proteins to use for the test dataset. Default: \code{30}.}

\item{correlation.method}{String indicating the method to use for the correlations. One among: 'pearson', 'spearman'. Default: \code{"pearson"}.}

\item{sample.group.column}{String indicating the ID of any column of the metadata table. This will be used to introduce the same frequencies of n-missing values for a protein and therefore not introducing the NAs completely at random in the dataset. Default: \code{NULL}, NAs are assigned randomly (same percentage of NAs present in the original table).}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'randomized', 'random'. Default: \code{"randomized"}.}

\item{masking}{String indicating how the values to hide are chosen. With \code{"MCAR"} every measured value has the same probability of being masked; with \code{"intensity"} the probability follows a logistic dropout curve estimated on the data, so that the masked values are drawn preferentially from the low-intensity range. One among: 'MCAR', 'intensity'. Default: \code{"MCAR"}.}

\item{max.NA.per.row}{Numeric value indicating the maximum number of missing values that a protein can already have to be usable in the test dataset. The RMSE is computed only on the artificially masked cells, hence a protein does not need to be complete: it is enough that the masked value was measured. Raising this threshold recovers the low-abundance proteins, which are rarely complete. Default: \code{0} (only complete proteins, as in the previous versions).}

\item{max.NA.fraction}{Numeric value between 0 and 1 indicating the maximum fraction of missing values that a protein can reach after the masking. It prevents the test dataset from becoming unusable for the algorithms that need a minimum number of measured values per protein. Ignored when \code{max.NA.per.row = 0}. Default: \code{0.5}.}

\item{dropout.by}{String indicating the level at which the dropout curves are estimated when \code{masking = "intensity"}. With \code{"group"} one curve is estimated for each level of \code{sample.group.column}, with \code{"sample"} one curve is estimated for each sample. One among: 'group', 'sample'. Default: \code{"group"}.}

\item{run.missForest}{Logical values indicating whether the test for the \code{missForest} imputation should be performed. Default: \code{TRUE}.}

\item{run.kNN}{Logical values indicating whether the test for the \code{kNN} imputation method should be performed. Default: \code{TRUE}.}

\item{run.tkNN}{Logical values indicating whether the test for the \code{tkNN} imputation method should be performed. Default: \code{TRUE}.}

\item{run.corkNN}{Logical values indicating whether the test for the \code{corkNN} imputation method should be performed. Default: \code{TRUE}.}

\item{run.LLS}{Logical values indicating whether the test for the \code{LLS} imputation method should be performed. Default: \code{TRUE}.}

\item{run.SVD}{Logical values indicating whether the test for the \code{SVD} imputation method should be performed. Default: \code{TRUE}.}

\item{run.BPCA}{Logical values indicating whether the test for the \code{BPCA} imputation method should be performed. Default: \code{TRUE}.}

\item{run.PPCA}{Logical values indicating whether the test for the \code{PPCA} imputation method should be performed. Default: \code{TRUE}.}

\item{run.RegImpute}{Logical values indicating whether the test for the \code{RegImpute} imputation method should be performed. Default: \code{TRUE}.}

\item{missForest.max.iterations}{Max number of iterations for the \code{missForest} algorithm. Default: \code{100}.}

\item{missForest.variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{missForest.cores}{Number of cores used to run the \code{missForest} algorithm. If \code{missForest.cores} is 1 (or lower), the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{missForest.parallel.mode}. Default: \code{1}.}

\item{missForest.parallel.mode}{Define the mode to use for the parallelization, ignored when \code{cores} is more than 1. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{kNN.n.nearest.neighbours}{Numeric value indicating the number of nearest neighbors to use to perform the \code{kNN} imputation. Default: \code{10}.}

\item{LLS.k}{Cluster size, this is the number of similar genes used for regression. Default: \code{2}.}

\item{pcaMethods.nPCs.to.test}{Numeric value indicating the number of Principal Components to test in order to find the optimal number of PCs to used in the imputation methods from the \code{pcaMethods} package. This includes: 'LLS', 'SVD' (a.k.a 'svdImpute'), 'BPCA-pcaMethods', and 'PPCA'. Default: \code{5}.}

\item{RegImpute.max.iterations}{Numeric value indicating the number of maximum iteration for the imputation with \code{RegImpute} (from \code{DreamAI}). Default: \code{10}.}

\item{RegImpute.fillmethod}{String identifying the fill method to be used in the \code{RegImpute} method (from\code{DreamAI}). One among \code{"row_mean"} and \code{"zeros"}. Default: \code{"row_mean"}. It throws an warning if \code{"row_median"} is used.}

\item{normalize.color.bar}{Logical value indicating whether the color bar limits for the residuals in the correlation plots should be normalized among the methods. Default: \code{TRUE}, the residual color bar absolute maximum is set to the max of all the residuals identified in all the methods,}

\item{low.residual.color}{String indicating any R-supported color that must be used for the negative values of the residuals color bar in the correlation plots. Default: \code{"firebrick"}.}

\item{zero.residual.color}{String indicating any R-supported color that must be used for the null residuals (zero, mid gradient color) color bar in the correlation plots. Default: \code{"white"}.}

\item{high.residual.color}{String indicating any R-supported color that must be used for the positive values of the residuals color bar in the correlation plots. Default: \code{"steelblue4"}.}

\item{seed}{Numeric value indicating the seed to use for the randomization. Default: \code{NULL}, automatically generated (saved in the \code{seed} slot in the final object).}

\item{verbose}{Logical valued indicating whether processing messages should be printed. Default: \code{FALSE}.}
}
\value{
A \code{DEprot.RMSE} object.
}
\description{
Function that computes the Root Mean Squared Error (RMSE) for all the 4 possible imputation algorithms: \href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://rdrr.io/cran/VIM/man/kNN.html}{kNN}, \href{https://rdrr.io/bioc/pcaMethods/man/llsImpute.html}{LLS}, \href{https://rdrr.io/bioc/pcaMethods/man/pca.html}{SVD}.
A new dataset containing only proteins with known values is created, a certain percentage of NAs is then manually introduced (with the possibility to respect the "pattern" of the missing values). This percentage is equivalent to the percentage of missing values in the original data set. At the end the imputed values in the new data set are compared with the measured (expected) ones.
}
\examples{
comparison <- compare.imp.methods(DEprot.object = DEprot::test.toolbox$dpo.norm,
                                  percentage.test = 100,
                                  sample.group.column = "combined.id",
                                  which.data = "normalized",
                                  run.missForest = FALSE,
                                  run.BPCA = TRUE,
                                  run.PPCA = TRUE,
                                  run.kNN = FALSE,
                                  run.tkNN = FALSE,
                                  run.corkNN = TRUE,
                                  run.LLS = TRUE,
                                  run.SVD = FALSE,
                                  run.RegImpute = FALSE)

summary(comparison)

comparison

}
\seealso{
\href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://cran.r-project.org/web/packages/VIM/index.html}{VIM}, \href{https://www.bioconductor.org/packages/release/bioc/html/pcaMethods.html}{pcaMethods} R-packages.
}
