% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.imp.methods.R
\name{compare.imp.methods}
\alias{compare.imp.methods}
\title{compare.imp.methods}
\usage{
compare.imp.methods(
  DEprot.object,
  percentage.test = 30,
  use.normalized.data = TRUE,
  missForest.max.iterations = 100,
  missForest.variable.wise.OOBerror = TRUE,
  missForest.cores = 1,
  missForest.parallel.mode = "variables",
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts} or \link{load.counts2}.}

\item{percentage.test}{Numeric value between 0 (excluded) and 100 indicating the percentage of proteins to use for the test dataset. Default: \code{30}.}

\item{use.normalized.data}{Logical value indicating whether the imputation should be performed based on the rationalized data. Default: \code{TRUE}.}

\item{missForest.max.iterations}{Max number of iterations for the missForest algorithm. Default: \code{100}.}

\item{missForest.variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{missForest.cores}{Number of cores used to run the missForest algorithm. If \code{missForest.cores} is 1 (or lower), the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{missForest.parallel.mode}. Default: \code{1}.}

\item{missForest.parallel.mode}{Define the mode to use for the parallelization, ignored when \code{cores} is more than 1. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{seed}{Numeric value indicating the seed to use for the randomization. Default: \code{NULL}, automatically generated (saved in the \code{seed} slot in the final object).}

\item{verbose}{Logical valued indicating whether processing messages should be printed. Default: \code{FALSE}.}

\item{overwrite.imputation}{Logical value to indicate whether, in the case already available, the table of imputed counts should be overwritten. Default: \code{FALSE}.}
}
\value{
A \code{DEprot.RMSE} object.
}
\description{
Function that computes the Root Mean Squared Error (RMSE) for all the 4 possibile imputation algorithms: \href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://rdrr.io/cran/VIM/man/kNN.html}{kNN}, \href{https://rdrr.io/bioc/pcaMethods/man/llsImpute.html}{LLS}, \href{https://rdrr.io/bioc/pcaMethods/man/pca.html}{SVD}.
A new dataset containing only proteins with known values is created, a certain percentage of NAs is then manually introduced. This percentage is equivalent to the precentage of missing values in the original data set. At the end the imputed values in the new data set are compared with the measured (expected) ones.
}
\seealso{
\href{https://www.rdocumentation.org/packages/missForest/}{missForest package}, href{https://cran.r-project.org/web/packages/VIM/index.html}{VIM package}, href{https://www.bioconductor.org/packages/release/bioc/html/pcaMethods.html}{pcaMethods package}.
}
