% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.imp.methods.R
\name{compare.imp.methods}
\alias{compare.imp.methods}
\title{compare.imp.methods}
\usage{
compare.imp.methods(
  DEprot.object,
  percentage.test = 30,
  sample.group.column = NULL,
  use.normalized.data = TRUE,
  run.missForest = TRUE,
  run.kNN = TRUE,
  run.LLS = TRUE,
  run.SVD = TRUE,
  missForest.max.iterations = 100,
  missForest.variable.wise.OOBerror = TRUE,
  missForest.cores = 1,
  missForest.parallel.mode = "variables",
  kNN.n.nearest.neighbours = 10,
  LLS.k = 2,
  normalize.color.bar = TRUE,
  low.residual.color = "firebrick",
  zero.residual.color = "white",
  high.residual.color = "steelblue4",
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts} or \link{load.counts2}.}

\item{percentage.test}{Numeric value between 0 (excluded) and 100 indicating the percentage of proteins to use for the test dataset. Default: \code{30}.}

\item{sample.group.column}{String indicating the ID of any column of the metadata table. This will be used to introduce the same frequencies of n-missing values for a protein and therefore not introducing the NAs completely at random in the dataset. Default: \code{NULL}, NAs are assigned randomly (same percentage of NAs present in the original table).}

\item{use.normalized.data}{Logical value indicating whether the imputation should be performed based on the rationalized data. Default: \code{TRUE}.}

\item{run.missForest}{Logical values indicating whether the test for the \code{missForest} imputation should be performed. Default: \code{TRUE}.}

\item{run.kNN}{Logical values indicating whether the test for the \code{kNN} imputation should be performed. Default: \code{TRUE}.}

\item{run.LLS}{Logical values indicating whether the test for the \code{LLS} imputation should be performed. Default: \code{TRUE}.}

\item{run.SVD}{Logical values indicating whether the test for the \code{SVD} imputation should be performed. Default: \code{TRUE}.}

\item{missForest.max.iterations}{Max number of iterations for the \code{missForest} algorithm. Default: \code{100}.}

\item{missForest.variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{missForest.cores}{Number of cores used to run the \code{missForest} algorithm. If \code{missForest.cores} is 1 (or lower), the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{missForest.parallel.mode}. Default: \code{1}.}

\item{missForest.parallel.mode}{Define the mode to use for the parallelization, ignored when \code{cores} is more than 1. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{kNN.n.nearest.neighbours}{Numeric value indicating the number of nearest neighbors to use to perform the \code{kNN} imputation. Default: \code{10}.}

\item{LLS.k}{Cluster size, this is the number of similar genes used for regression. Default: \code{2}.}

\item{normalize.color.bar}{Logical value indicating whether the color bar limits for the residuals in the correlation plots should be normalized among the methods. Default: \code{TRUE}, the residual color bar absolute maximum is set to the max of all the residuals identified in all the methods,}

\item{low.residual.color}{String indicating any R-supported color that must be used for the negative values of the residuals color bar in the correlation plots. Default: \code{"firebrick"}.}

\item{zero.residual.color}{String indicating any R-supported color that must be used for the null residuals (zero, mid gradient color) color bar in the correlation plots. Default: \code{"white"}.}

\item{high.residual.color}{String indicating any R-supported color that must be used for the positive values of the residuals color bar in the correlation plots. Default: \code{"steelblue4"}.}

\item{seed}{Numeric value indicating the seed to use for the randomization. Default: \code{NULL}, automatically generated (saved in the \code{seed} slot in the final object).}

\item{verbose}{Logical valued indicating whether processing messages should be printed. Default: \code{FALSE}.}

\item{overwrite.imputation}{Logical value to indicate whether, in the case already available, the table of imputed counts should be overwritten. Default: \code{FALSE}.}
}
\value{
A \code{DEprot.RMSE} object.
}
\description{
Function that computes the Root Mean Squared Error (RMSE) for all the 4 possibile imputation algorithms: \href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://rdrr.io/cran/VIM/man/kNN.html}{kNN}, \href{https://rdrr.io/bioc/pcaMethods/man/llsImpute.html}{LLS}, \href{https://rdrr.io/bioc/pcaMethods/man/pca.html}{SVD}.
A new dataset containing only proteins with known values is created, a certain percentage of NAs is then manually introduced (with the possibility to respect the "pattern" of the missing values). This percentage is equivalent to the percentage of missing values in the original data set. At the end the imputed values in the new data set are compared with the measured (expected) ones.
}
\seealso{
\href{https://www.rdocumentation.org/packages/missForest/}{missForest package}, \href{https://cran.r-project.org/web/packages/VIM/index.html}{VIM package}, \href{https://www.bioconductor.org/packages/release/bioc/html/pcaMethods.html}{pcaMethods package}.
}
