% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.ranking.R
\name{compare.ranking}
\alias{compare.ranking}
\title{compare.ranking}
\usage{
\method{compare}{ranking}(
  DEprot.analyses.object,
  contrast,
  color.up = "indianred",
  color.down = "steelblue",
  regression.line.color = "purple"
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Number indicating the position of the contrast to use for the plotting.}

\item{color.up}{String indicating any R-supported color that will be used for dots with a positive fold change. Default: \code{"indianred"}.}

\item{color.down}{String indicating any R-supported color that will be used for dots with a negative fold change. Default: \code{"steelblue"}.}

\item{regression.line.color}{String indicating any R-supported color to use for the regression line and error. Default: \code{"purple"}.}
}
\value{
Two ggplots combined in an object of class \code{patchwork}.
}
\description{
Compare the two possible methods for the gene ranking used to perform Gene Set Enrichment Analyses (GSEA).
}
\examples{
ranking <- compare.ranking(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma, contrast = 1)


}
