% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{DEprot-dollar-methods}
\alias{DEprot-dollar-methods}
\title{Slot access for DEprot classes}
\arguments{
\item{x}{An object of one of the DEprot classes: \code{DEprot}, \code{DEprot.analyses},
\code{DEprot.PCA}, \code{DEprot.PCoA}, \code{DEprot.sPLSDA}, \code{DEprot.correlation}, \code{DEprot.counts.heatmap},
\code{DEprot.contrast.heatmap}, \code{DEprot.upset}, \code{DEprot.normality}, \code{DEprot.pvalues},
\code{DEprot.enrichResult}, \code{DEprot.RMSE}, \code{DEprot.SAINTq}, \code{DEprot.missingness},
\code{DEprot.outliers}, \code{DEprot.timecourse}, \code{DEprot.timecourse.enrichment}, \code{DEprot.power}.}

\item{name}{Name of the slot to access or replace.}

\item{value}{The value to assign to the slot (only for \code{$<-}).}
}
\value{
For \code{$}, the content of the corresponding slot. For \code{$<-}, the object with the updated slot.
}
\description{
Instance-aware slot access and replacement for all the DEprot S4 classes.
These methods allow the slots of a DEprot object to be accessed and modified with the
\code{$} operator (in addition to the standard \code{@} operator), enabling slot-name
auto-completion in interactive sessions.
}
\author{
Sebastian Gregoricchio
}
