% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.analyses.limma.R
\name{diff.analyses.limma}
\alias{diff.analyses.limma}
\title{diff.analyses.limma}
\usage{
\method{diff}{analyses.limma}(
  DEprot.object,
  contrast.list,
  include.rep.model = FALSE,
  replicate.column = NULL,
  linear.FC.th = 2,
  linear.FC.unresp.range = c(1/1.1, 1.1),
  padj.th = 0.05,
  padj.method = "BH",
  fitting.method = "ls",
  up.color = "indianred",
  down.color = "steelblue",
  unresponsive.color = "purple",
  null.color = "gray",
  which.data = "imputed",
  overwrite.analyses = FALSE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{contrast.list}{List of 3-elements vectors indicating (in order): metadata_column, variable_1, variable_2.}

\item{include.rep.model}{Logical value indicating whether the model should include the replicate information. Requires a 'replicate.column'. Default: \code{FALSE}.}

\item{replicate.column}{String indicating the name of a column from the metadata table in which are stored the replicate IDs. This column is used only if \code{paired.test = TRUE}. Default: \code{NULL}.}

\item{linear.FC.th}{Number indicating the (absolute) fold change threshold (linear scale) to use to define differential proteins. Default: \code{2}.}

\item{linear.FC.unresp.range}{A numeric 2-elements vector indicating the range (linear scale) used to define the unresponsive fold changes. Default: \code{c(1/1.1, 1.1)}.}

\item{padj.th}{Numeric value indicating the p.adjusted threshold to apply to the differential analyses. Default: \code{0.05}.}

\item{padj.method}{String indicating the method to use to correct the p-values. One among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: \code{BH}.}

\item{fitting.method}{String indicating the method that limma should use to fit the model. Options: "ls" for least squares or "robust" for robust regression. Default: \code{"ls"} (least squares).}

\item{up.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"indianred"}.}

\item{down.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"steelblue"}.}

\item{unresponsive.color}{String indicating the color to use for unresponsive proteins in the plots. Default: \code{"purple"}.}

\item{null.color}{String indicating the color to use for null proteins in the plots. Default: \code{"gray"}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{overwrite.analyses}{Logical value to indicate whether overwrite analyses already generated. Default: \code{FALSE}.}
}
\description{
Allows for the computation of differential analyses using \href{https://bioconductor.org/packages/release/bioc/html/limma.html}{\code{limma}}. Includes means, Fold Changes, and p-values.
}
\author{
Sebastian Gregoricchio
}
