% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.analyses.R
\name{diff.analyses}
\alias{diff.analyses}
\title{diff.analyses}
\usage{
\method{diff}{analyses}(
  DEprot.object,
  contrast.list,
  linear.FC.th = 2,
  linear.FC.unresp.range = c(1/1.1, 1.1),
  padj.th = 0.05,
  padj.method = "BH",
  stat.test = "t.test",
  up.color = "indianred",
  down.color = "steelblue",
  unresponsive.color = "purple",
  null.color = "gray",
  which.data = "imputed",
  overwrite.analyses = FALSE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{contrast.list}{List of 3-elements vectors indicating (in order): metadata_column, variable_1, variable_2.}

\item{linear.FC.th}{Number indicating the (absolute) fold change threshold (linear scale) to use to define differential proteins. Default: \code{2}.}

\item{linear.FC.unresp.range}{A numeric 2-elements vector indicating the range (linear scale) used to define the unresponsive fold changes. Default: \code{c(1/1.1, 1.1)}.}

\item{padj.th}{Numeric value indicating the p.adjusted threshold to apply to the differential analyses. Default: \code{0.05}.}

\item{padj.method}{String indicating the method to use to correct the p-values. One among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: \code{BH}.}

\item{stat.test}{String indicating the type of statistic test to use. One among: "t-test" and "wilcoxon". Default: \code{"t.test"}.}

\item{up.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"indianred"}.}

\item{down.color}{String indicating the color to use for up-regulated proteins in the plots. Default: \code{"steelblue"}.}

\item{unresponsive.color}{String indicating the color to use for unresponsive proteins in the plots. Default: \code{"purple"}.}

\item{null.color}{String indicating the color to use for null proteins in the plots. Default: \code{"gray"}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{overwrite.analyses}{Logical value to indicate whether overwrite analyses already generated. Default: \code{FALSE}.}
}
\description{
Allows for the computation of differential analyses. Includes means, Fold Changes, and pvalues.
}
