% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.corum.reshape}
\alias{.corum.reshape}
\title{.corum.reshape}
\usage{
.corum.reshape(corum_table)
}
\arguments{
\item{corum_table}{Either the path to the CORUM database file or a data.frame with the same structure.}
}
\value{
Data.frame with 4 columns: complex.id, complex.name, organism, protein.members.
}
\description{
Reshapes the CORUM database downloads files for gene entrenchment analyses
}
\keyword{internal}
