% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.analyses.R
\name{export.analyses}
\alias{export.analyses}
\title{export.analyses}
\usage{
export.analyses(
  DEprot.analyses.object,
  output.folder = "./export",
  contrast.subset = NULL,
  verbose = TRUE
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{output.folder}{Path (string) to the output directory. If the folder is not already existing it will be created. Default: \code{"./export"}.}

\item{contrast.subset}{Numeric vector indicating the contrasts to use. Default: \code{NULL} (all contrasts are exported).}

\item{verbose}{Logical value to indicate whether progress messages should be printed. Default: \code{TRUE}.}
}
\value{
Nothing is returned, data are exported locally.
}
\description{
Export all the results from a DEprot.analyses object into a user-defined folder.
}
\examples{
\dontrun{
  # Export data for all contrasts
  DEprot::export.analyses(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
                         output.folder = out_dir,
                          verbose = TRUE)


  # Export data for only a subset of contrasts
  DEprot::export.analyses(DEprot.analyses.object = dpo_analyses,
                          contrast.subset = c(1:3,6),
                          output.folder = out_dir,
                          verbose = TRUE)
}

}
\author{
Sebastian Gregoricchio
}
