% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.boxplot.R
\name{expression.boxplot}
\alias{expression.boxplot}
\title{expression.boxplot}
\usage{
expression.boxplot(
  DEprot.object,
  protein.id,
  which.data = "imputed",
  sample.subset = NULL,
  shape.column = NULL,
  group.by.metadata.column = "column.id",
  group.levels = NULL,
  scale.expression = FALSE,
  x.label.angle = 30
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{protein.id}{String indicating a protein for which plot the expression. The identifier must correspond to the full row.name of the counts table (equivalent to the \code{prot.id} column of the fold change table of \code{DEprot.analyses} object).}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{sample.subset}{Character vector indicating a subset of samples to display. The identifiers must correspond to a IDs in the \code{column.id} column of the object's metadata. Default: \code{NULL} (all samples are shown).}

\item{shape.column}{String indicating a column from the metadata table. This column will be used as factor for the shape of the points on the boxplot. Default: \code{NULL}: no different shapes.}

\item{group.by.metadata.column}{String indicating a column from the metadata table. This column will be used to define sample groups, and for each group it will be computed a mean of the counts. Default: \code{"column.id"} (no groups).}

\item{group.levels}{Ordered string vector indicating the order to use for the groups. Default: \code{NULL}, counts table order will be applied}

\item{scale.expression}{Logic value indicating whether Z-scores should be computed. Default: \code{FALSE} (no scaling).}

\item{x.label.angle}{Numeric value indicating the rotation angle to use for the x-axis labels. Default: \code{30}.}
}
\value{
A boxplot of class ggplot2.
}
\description{
Plots a boxplot of the expression of a specific protein. Samples can be groups depending on a metadata column.
}
\author{
Sebastian Gregoricchio
}
