% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.proteins.R
\name{filter.proteins}
\alias{filter.proteins}
\title{filter.proteins}
\usage{
\method{filter}{proteins}(DEprot.object, proteins, mode = "keep")
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{proteins}{A string vector indicating a list of proteins to be use as filtering criteria. It must correspond to the counts rown names and 'prot.id' column of the analyses object.}

\item{mode}{A string indicating the mode to use for the filtering. If 'keep' the proteins are selected and kept, while if 'remove' the proteins are discarded. Default: \code{"keep"}.}
}
\value{
An object of class \code{DEprot} (S4 vector).
}
\description{
Function that allows for removing or keeping specific proteins from any \code{DEprot} object class.
}
