% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.enrichment.R
\name{geneset.enrichment}
\alias{geneset.enrichment}
\title{geneset.enrichment}
\usage{
geneset.enrichment(
  DEprot.analyses.object,
  contrast,
  TERM2GENE,
  enrichment.type,
  gsea.rank.method = "foldchange",
  diff.status.category = NULL,
  gsub.pattern.prot.id = NULL,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  dotplot.n = 10
)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Number indicating the position of the contrast to use for the plotting.}

\item{TERM2GENE}{Data.frame containing two columns 'gs_name' (IDs of the gene sets) and 'gene_symbol' (indicating the gene IDs). No default.}

\item{enrichment.type}{String indicating the type of analyses to perform. One among: GSEA, ORA. No default.}

\item{gsea.rank.method}{String indicating the type of gene ranking to use for GSEA analyses. Possible options: \code{"foldchange"} (log2FC value of the contrast), \code{"correlation"} (spearman's correlation coefficient of the imputed counts between the two groups in the contrast). Default: \code{"foldchange"}.}

\item{diff.status.category}{String indicating a diff.status among the ones present in the results table of the specific contrast. Used only one 'ORA' is performed. Default: \code{NULL}.}

\item{gsub.pattern.prot.id}{String indicating a pattern to be passed to gsub and to remove from the prot.id. Default: \code{NULL} (non changes in the IDs).}

\item{pvalueCutoff}{Numeric value indicating the adjusted pvalue cutoff on enrichment tests to report. Default: \code{0.05}.}

\item{qvalueCutoff}{Numeric value indicating the qvalue cutoff on enrichment tests to report as significant (only for ORA). Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported. Default: \code{0.05}.}

\item{pAdjustMethod}{String indicating the method to use for the p-value adjustment. One mong "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: \code{"BH"}.}

\item{dotplot.n}{Numeric value indicating the maximum number of categories to plot in the dotplot. Default: \code{10}.}
}
\value{
An object of class \code{DEprot.enrichResult}.
}
\description{
Perform Gene Set Enrichment Analyses (GSEA) or OverRepresentation Analyses (ORA) on a proteomics differential analyses.
}
