% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metadata.R
\name{get.metadata}
\alias{get.metadata}
\title{get.metadata}
\usage{
get.metadata(DEprot.object)
}
\arguments{
\item{DEprot.object}{Any object of class \code{DEprot}.}
}
\value{
Data.frame corresponding to the metadata of the provided object.
}
\description{
Function to extract the metadata from a \code{DEprot} object.
}
