% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.results.R
\name{get.results}
\alias{get.results}
\title{get.results}
\usage{
get.results(DEprot.analyses.object, contrast = 1)
}
\arguments{
\item{DEprot.analyses.object}{An object of class \code{DEprot.analyses}.}

\item{contrast}{Number indicating the position of the contrast to use for the plotting.}
}
\value{
A data frame.
}
\description{
Simplifies the access to the differential expression results table
}
\examples{
get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)


}
\author{
Sebastian Gregoricchio
}
