% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.batches.R
\name{harmonize.batches}
\alias{harmonize.batches}
\title{harmonize.batches}
\usage{
harmonize.batches(DEprot.object, batch.column = "batch", cores = 1)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}. Notice that this function works only with "normalized".}

\item{batch.column}{String indicating the name of a column from the metadata table in which are stored the replicate IDs. This column is used only if \code{paired.test = TRUE}. Default: \code{NULL}.}

\item{cores}{Numeric value indicating the number of cores to use for the batch correction. Default: \code{1}.}
}
\value{
An object of class \code{DEprot}
}
\description{
The function uses the package \href{https://www.bioconductor.org/packages/release/bioc/html/HarmonizR.html}{\code{HarmonizR}} to correct batch effects in a combined table of multiple experiments.
}
\examples{
# Add batch column to the DEprot object
dpo = DEprot::test.toolbox$dpo.raw
dpo@metadata$batch = c(rep("A",6), rep("B",6))

# Correct batch effects
dpo <- harmonize.batches(DEprot.object = dpo,
                         batch.column = "batch",
                         cores = 1)


}
\author{
Sebastian Gregoricchio
}
