% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.counts.R
\name{heatmap.counts}
\alias{heatmap.counts}
\title{heatmap.counts}
\usage{
heatmap.counts(
  DEprot.object,
  which.data = "imputed",
  contrast = NULL,
  top.n = NULL,
  sample.subset = NULL,
  protein.subset = NULL,
  group.by.metadata.column = NULL,
  scale = NULL,
  clust.rows = TRUE,
  clust.columns = TRUE,
  distance.method = "euclidean",
  clustering.method = "complete",
  palette = RColorBrewer::brewer.pal(n = 9, name = "Blues"),
  high.color = "firebrick",
  low.color = "#2166AC",
  mid.color = "white",
  na.color = "gray",
  color.limits = c(NA, NA),
  cell.border.color = NA,
  cell.border.width = 0.5,
  show.protein.names = FALSE,
  protein.names.pattern = NULL,
  title = NULL,
  use.uncorrected.pvalue = FALSE
)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot} or \code{DEprot.analyses}.}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}

\item{contrast}{Numeric vector indicating the position of the contrast to use for the plotting. Only differential proteins in this contrast will be shown. Option available only for an object of class \code{DEprot.analyses}. Default: \code{NULL} (non differential protein selection).}

\item{top.n}{Numeric value indicated the top differentially expressed proteins to consider for the contrast selected. The rank is based on the product of log2Fc and -log10Padj. Option available only for an object of class \code{DEprot.analyses}. Default: \code{NULL} (all differential proteins of that contrast).}

\item{sample.subset}{Character vector indicating a subset of samples to display. The identifiers must correspond to a IDs in the \code{column.id} column of the object's metadata. Default: \code{NULL} (all samples are shown).}

\item{protein.subset}{Character vector indicating a subset of proteins to display. The identifiers must correspond to the full row.names of the counts table (equivalent to the \code{prot.id} column of the fold change table of \code{DEprot.analyses} object). This options is can be used in combination with \code{contrast} and \code{top.n}. Default: \code{NULL} (all proteins are shown).}

\item{group.by.metadata.column}{String indicating a column from the metadata table. This column will be used to define sample groups, and for each group it will be computed a mean of the counts. Default: \code{NULL} (no groups).}

\item{scale}{String indicating whether Z-scores should be computed. Possible choices: "row" or "column". Default: \code{NULL} (no scaling).}

\item{clust.rows}{Logical value indicating whether heatmap rows (proteins) should be clustered. Default: \code{TRUE}.}

\item{clust.columns}{Logical value indicating whether heatmap columns (samples or groups) should be clustered. Default: \code{TRUE}.}

\item{distance.method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given. Default: \code{"euclidean"}.}

\item{clustering.method}{The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (UPGMA), "mcquitty" (WPGMA), "median" (WPGMC) or "centroid" (UPGMC). Default: \code{complete"}.}

\item{palette}{List of colors to use for the color gradient of the heatmap. This parameters is used when raw-counts are plotted. For scaled data see the high, low and mid parameters. Default: \code{RColorBrewer::brewer.pal(n = 9, name = "Blues")}.}

\item{high.color}{String indicating the color to use for positive Z-score for scaled data. Default: \code{"indianred"}.}

\item{low.color}{String indicating the color to use for negative Z-score for scaled data. Default: \code{"#2166AC"} (blue).}

\item{mid.color}{String indicating the color to use for the 0 Z-score value for scaled data. Default: \code{"white"}.}

\item{na.color}{String indicating the color to use for the NA values in the heatmap. Default: \code{"gray"}.}

\item{color.limits}{Numeric vector of length 2 indicating lower and upper limit of the color scale values. Default: \code{c(NA,NA)}, automatic limits applied.}

\item{cell.border.color}{String indicating the color to use for the individual cells border. Default: \code{NA} (no border).}

\item{cell.border.width}{Numeric value indicating the width of the cell border line. Ignored when \code{cell.border.color = NA}. Default: \code{0.5}.}

\item{show.protein.names}{Logical value to indicate whether the protein names should be displayed. Default: \code{FALSE}.}

\item{protein.names.pattern}{Character indicating a regular expression to remove from the protein IDs. Default" \code{NULL}, no alterations in the protein IDs.}

\item{title}{String indicating the title to use, markdown formatting supported. Default: \code{NULL} (automatic title).}

\item{use.uncorrected.pvalue}{Logical value indicating whether it should be used the normal p-value instead of the adjusted one (differential proteins numbers are recomputed). Default: \code{FALSE}, padj is used.}
}
\value{
A \code{DEprot.contrast.heatmap} object, which contains the heatmap and the hclust object used to order the rows.
}
\description{
Plots an heatmap of the counts (raw, normalized, or imputed). It is possible to perform a scaling (z-score) by row or by column.
}
