% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.counts.R
\name{impute.counts}
\alias{impute.counts}
\title{impute.counts}
\usage{
impute.counts(
  DEprot.object,
  max.iterations = 100,
  variable.wise.OOBerror = TRUE,
  use.normalized.data = TRUE,
  overwrite.imputation = FALSE,
  cores = 1,
  parallel.mode = "variables",
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{max.iterations}{Max number of iterations for the missForest algorithm. Default: \code{100}.}

\item{variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{use.normalized.data}{Logical value indicating whether the imputation should be performed based on the rationalized data. Default: \code{TRUE}.}

\item{overwrite.imputation}{Logical value to indicate whether, in the case already available, the table of imputed counts should be overwritten. Default: \code{FALSE}.}

\item{cores}{Number of cores used to run the missForest algorithm. If \code{cores} is 1 (or lower), the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{parallel.mode}. Default: \code{1}.}

\item{parallel.mode}{Define the mode to use for the parallelization, ignored when \code{cores} is more than 1. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{verbose}{Logical valued indicating whether processing messages should be printed. Default: \code{FALSE}.}
}
\value{
A \code{DEprot} object. The boxplot showing the distribution of the protein intensity is remade and added to the slot (\code{boxplot.imputed}). A list with parameters and other info about the imputation is added as well in the \code{imputation} slot.
}
\description{
Function that allows for the imputation of missing values using the \href{https://www.rdocumentation.org/packages/missForest/}{missForest} algorithm.
}
\seealso{
\href{https://www.rdocumentation.org/packages/missForest/}{missForest package}.
}
