% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.counts.R
\name{impute.counts}
\alias{impute.counts}
\title{impute.counts}
\usage{
impute.counts(
  DEprot.object,
  method = "missForest",
  use.normalized.data = TRUE,
  overwrite.imputation = FALSE,
  missForest.max.iterations = 100,
  missForest.variable.wise.OOBerror = TRUE,
  missForest.cores = 1,
  missForest.parallel.mode = "variables",
  kNN.n.nearest.neighbours = 10,
  LLS.k = 2,
  verbose = FALSE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{method}{String indicating the imputation method to use. One among: 'missForest', 'kNN', 'LLS', 'SVD'. Default: \code{"missForest"}.}

\item{use.normalized.data}{Logical value indicating whether the imputation should be performed based on the rationalized data. Default: \code{TRUE}.}

\item{overwrite.imputation}{Logical value to indicate whether, in the case already available, the table of imputed counts should be overwritten. Default: \code{FALSE}.}

\item{missForest.max.iterations}{Max number of iterations for the missForest algorithm. Default: \code{100}.}

\item{missForest.variable.wise.OOBerror}{Logical value to define whether the OOB error is returned for each variable separately. Default: \code{TRUE}.}

\item{missForest.cores}{Number of cores used to run the \code{missForest} algorithm. If \code{missForest.cores} is 1 (or lower), the imputation will be run in parallel. Two modes are possible and can be defined by the parameter \code{missForest.parallel.mode}. Default: \code{1}.}

\item{missForest.parallel.mode}{Define the mode to use for the parallelization, ignored when \code{cores} is more than 1. One among: 'variables', 'forests'. Default: \code{"variables"}. See also the documentation of the \href{https://www.rdocumentation.org/packages/missForest/versions/1.5/topics/missForest}{missForest function}.}

\item{kNN.n.nearest.neighbours}{Numeric value indicating the number of nearest neighbors to use to perform the \code{kNN} imputation. Default: \code{10}.}

\item{LLS.k}{Cluster size, this is the number of similar genes used for regression. Default: \code{2}.}

\item{verbose}{Logical valued indicating whether processing messages should be printed. Default: \code{FALSE}.}
}
\value{
A \code{DEprot} object. The boxplot showing the distribution of the protein intensity is remade and added to the slot (\code{boxplot.imputed}). A list with parameters and other info about the imputation is added as well in the \code{imputation} slot.
}
\description{
Function that allows for the imputation of missing values using 4 possible algorithms: \href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://rdrr.io/cran/VIM/man/kNN.html}{kNN}, \href{https://rdrr.io/bioc/pcaMethods/man/llsImpute.html}{LLS}, \href{https://rdrr.io/bioc/pcaMethods/man/pca.html}{SVD}.
}
\seealso{
\href{https://www.rdocumentation.org/packages/missForest/}{missForest}, \href{https://cran.r-project.org/web/packages/VIM/index.html}{VIM}, \href{https://www.bioconductor.org/packages/release/bioc/html/pcaMethods.html}{pcaMethods} R-packages.
}
