% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.counts.R
\name{load.counts}
\alias{load.counts}
\title{load.counts}
\usage{
load.counts(
  counts,
  metadata,
  log.base,
  imputation = NA,
  normalization.method = NA,
  column.id = "column.id"
)
}
\arguments{
\item{counts}{A data.frame or a matrix in which the rownames are the proteins and the columns the samples.}

\item{metadata}{A data.frame containing at least one column called \code{column.id} which corresponds to the colnames of \code{counts}. Any other column can be added and will correspond to a "feature"of each sample.}

\item{log.base}{Number indicating the base of the log used to transform the counts. If none transformation is applied, indicate the base \code{1}.}

\item{imputation}{A string indicating the imputation method used. If none, use the default value \code{NA}.}

\item{normalization.method}{String or list indicating the normalization method used. If none, use the default value \code{NA}.}

\item{column.id}{String indicating the name of the column to use as "column.id" from the metadata data.frame. This column must contain all the colnames of \code{counts}.}
}
\value{
A \code{DEprot} object (S4 vector).
}
\description{
Function used to generate a \code{DEprot} object starting from counts and metadata.
}
\examples{
# Raw counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2)

# Normalized counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2,
                   normalization.method = "MaxQuant automatic")

# Imputed counts
dpo <- load.counts(counts = DEprot::unimputed.counts,
                   metadata = DEprot::sample.config,
                   log.base = 2,
                   imputation = "missForest")



}
\author{
Sebastian Gregoricchio
}
