% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.counts2.R
\name{load.counts2}
\alias{load.counts2}
\title{load.counts2}
\usage{
load.counts2(
  counts,
  metadata,
  data.type,
  log.base,
  normalization.method = NA,
  imputation.method = NA,
  column.id = "column.id"
)
}
\arguments{
\item{counts}{A data.frame or a matrix in which the rownames are the proteins and the columns the samples.}

\item{metadata}{A data.frame containing at least one column called \code{column.id} which corresponds to the colnames of \code{counts}. Any other column can be added and will correspond to a "feature"of each sample.}

\item{data.type}{String indicating the type of data that are loaded. One among: 'raw', 'normalized', 'imputed'.}

\item{log.base}{Number indicating the base of the log used to transform the counts. If none transformation is applied, indicate the base \code{1}.}

\item{normalization.method}{String or list indicating the normalization method used. If none, use the default value \code{NA}.}

\item{imputation.method}{A string indicating the imputation method used. If none, use the default value \code{NA}.}

\item{column.id}{String indicating the name of the column to use as "column.id" from the metadata data.frame. This column must contain all the colnames of \code{counts}.}
}
\value{
A \code{DEprot} object (S4 vector).
}
\description{
Function used to generate a \code{DEprot} object starting from counts and metadata.
}
\author{
Sebastian Gregoricchio
}
