% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.PCA.R
\name{perform.PCA}
\alias{perform.PCA}
\title{perform.PCA}
\usage{
perform.PCA(DEprot.object, column.subset = NULL, use.normalized.data = TRUE)
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{column.subset}{String vector indicating the column names to keep in the counts table (or 'column.id' in the metdata table). Default: \code{NULL} (no subsetting).}

\item{use.normalized.data}{Logical value to indicate whether to use the normalized data or not. Default: \code{TRUE}.}
}
\value{
A \code{DEprot.PCA}, containing the PC values (\code{PCs}) and the importance summary (\code{importance}).
}
\description{
This function performs principal component analyses (PCA).
}
