% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.PCA.R
\name{perform.PCA}
\alias{perform.PCA}
\title{perform.PCA}
\usage{
perform.PCA(DEprot.object, sample.subset = NULL, which.data = "imputed")
}
\arguments{
\item{DEprot.object}{An object of class \code{DEprot}.}

\item{sample.subset}{String vector indicating the column names (samples) to keep in the counts table (the 'column.id' in the metadata table). Default: \code{NULL} (no subsetting).}

\item{which.data}{String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: \code{"imputed"}.}
}
\value{
A \code{DEprot.PCA}, containing the PC values (\code{PCs}) and the importance summary (\code{importance}).
}
\description{
This function performs principal component analyses (PCA).
}
