% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot,ANY-method}
\alias{plot,DEprot,ANY-method}
\title{DEprot plot-method}
\usage{
\S4method{plot}{DEprot,ANY}(x, y, ..., ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of class \code{DEprot}.}

\item{y}{Not used.}

\item{...}{Not used.}
}
\description{
DEprot plot-method
}
\keyword{internal}
