% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.analyses,ANY-method}
\alias{plot,DEprot.analyses,ANY-method}
\title{DEprot.analyses plot-method}
\usage{
\S4method{plot}{DEprot.analyses,ANY}(x, y, ..., plot.type = "volcano", ncol = NULL, nrow = NULL)
}
\arguments{
\item{x}{Object of class \code{DEprot.analyses}}

\item{y}{Not used.}

\item{...}{Not used.}
}
\description{
DEprot.analyses plot-method
}
\keyword{internal}
