% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.normality,ANY-method}
\alias{plot,DEprot.normality,ANY-method}
\title{DEprot.normality plot-method}
\usage{
\S4method{plot}{DEprot.normality,ANY}(x, y, ..., n.samples = NULL)
}
\arguments{
\item{x}{Object of class \code{DEprot.normality}}

\item{y}{Not used.}

\item{...}{Not used.}
}
\description{
DEprot.normality plot-method
}
\keyword{internal}
