% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.timecourse,ANY-method}
\alias{plot,DEprot.timecourse,ANY-method}
\title{DEprot.timecourse plot-method}
\usage{
\S4method{plot}{DEprot.timecourse,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{DEprot.timecourse}.}

\item{y}{Not used.}

\item{...}{Not used.}
}
\value{
The cluster-profile plot stored in the object.
}
\description{
DEprot.timecourse plot-method
}
