% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{plot,DEprot.timecourse,ANY-method}
\alias{plot,DEprot.timecourse,ANY-method}
\title{plot method for DEprot.timecourse objects}
\usage{
\S4method{plot}{DEprot.timecourse,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{DEprot.timecourse}.}

\item{y}{Not used.}

\item{...}{Not used.}
}
\value{
The cluster-profile plot stored in the object.
}
\description{
plot method for DEprot.timecourse objects
}
