% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{plot,DEprot.timecourse.enrichment,ANY-method}
\alias{plot,DEprot.timecourse.enrichment,ANY-method}
\title{plot method for DEprot.timecourse.enrichment objects}
\usage{
\S4method{plot}{DEprot.timecourse.enrichment,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{DEprot.timecourse.enrichment}.}

\item{y}{Not used.}

\item{...}{Not used.}
}
\value{
The dotplot stored in the object.
}
\description{
plot method for DEprot.timecourse.enrichment objects
}
