% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PC.cumulative.R
\name{plot.PC.cumulative}
\alias{plot.PC.cumulative}
\title{plot.PC.cumulative}
\usage{
plot.PC.cumulative(
  DEprot.PCA.object,
  bar.color = "steelblue",
  line.color = "navyblue",
  title = NULL
)
}
\arguments{
\item{DEprot.PCA.object}{An object of class \code{DEprot.PCA}.}

\item{bar.color}{String indicating the color to use for the bar fill. Default: \code{"steelblue"}.}

\item{line.color}{String indicating the color to use for the line and the dots of the cumulative curve. Default: \code{"navyblue"}.}

\item{title}{String indicating the title of the plot (markdown annotation supported).}
}
\value{
A barplot in ggplot format.
}
\description{
Function to plot the cumulative variance of all the principal components of a PCA.
}
\examples{
# Perform Principal Component Analyses (PCA)
pca <- perform.PCA(DEprot.object = DEprot::test.toolbox$dpo.imp)

# Plot cumulative variance barplot
plot.PC.cumulative(pca)

}
\seealso{
\link{perform.PCA}, \link{plot.PCoA.cumulative}
}
\author{
Sebastian Gregoricchio
}
