% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PC.scatter.123.R
\name{plot.PC.scatter.123}
\alias{plot.PC.scatter.123}
\title{plot.PC.scatter.123}
\usage{
\method{plot}{PC.scatter.123}(
  DEprot.PCA.object,
  color.column = "column.id",
  shape.column = NULL,
  label.column = NULL,
  dot.colors = NULL,
  title = NULL,
  plot.zero.line.y.12 = TRUE,
  plot.zero.line.x.12 = TRUE,
  plot.zero.line.y.23 = TRUE,
  plot.zero.line.x.23 = TRUE
)
}
\arguments{
\item{DEprot.PCA.object}{An object of class \code{DEprot.PCA.object}, as generated by \link{perform.PCA}.}

\item{color.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot colors. Default: \code{"column.id"} (each sample a color).}

\item{shape.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the \code{metadata} to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{dot.colors}{Color-vector indicating the colors to use for the points in the plot. If \code{NULL} (default) or the number of colors is lower than the required, automatic colors will be assign using the \code{rainbow} function. Default: \code{NULL}.}

\item{title}{String indicating a title to add to the plot (markdown syntax supported). Default: \code{NULL} (no)}

\item{plot.zero.line.y.12}{Logical value to indicate whether to plot two gray dashed lines in correspondence of y=0 in the PC1-vs-PC2 plot. Default: \code{TRUE}.}

\item{plot.zero.line.x.12}{Logical value to indicate whether to plot two gray dashed lines in correspondence of x=0 in the PC1-vs-PC2 plot. Default: \code{TRUE}.}

\item{plot.zero.line.y.23}{Logical value to indicate whether to plot two gray dashed lines in correspondence of y=0 in the PC2-vs-PC3 plot. Default: \code{TRUE}.}

\item{plot.zero.line.x.23}{Logical value to indicate whether to plot two gray dashed lines in correspondence of x=0 in the PC2-vs-PC3 plot. Default: \code{TRUE}.}
}
\value{
A patchwork object result of a combination of two ggplot objects.
}
\description{
Plot combined scatter of the PCs 1-2,2-3 from an object generated by \link{perform.PCA}.
}
\seealso{
\link{perform.PCA} and \link{plot.PC.scatter}
}
\author{
Sebastian Gregoricchio
}
