% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PC.scatter.R
\name{plot.PC.scatter}
\alias{plot.PC.scatter}
\title{plot.PC.scatter}
\usage{
\method{plot}{PC.scatter}(
  DEprot.PCA.object,
  PC.x = 2,
  PC.y = 1,
  color.column = "column.id",
  shape.column = NULL,
  label.column = NULL,
  plot.zero.line.x = TRUE,
  plot.zero.line.y = TRUE
)
}
\arguments{
\item{DEprot.PCA.object}{An object of class \code{DEprot.PCA.object}, as generated by \link{perform.PCA}.}

\item{PC.x}{Number indicating which Principal Component (PC) display on the x-axis. Default: \code{2} (PC2).}

\item{PC.y}{Number indicating which Principal Component (PC) display on the y-axis. Default: \code{1} (PC1).}

\item{color.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot colors. Default: \code{"column.id"} (each sample a color).}

\item{shape.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the \code{metadata} to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{plot.zero.line.x}{Logical value to indicate whether to plot two gray dashed line in correspondence of x=0. Default: \code{TRUE}.}

\item{plot.zero.line.y}{Logical value to indicate whether to plot two gray dashed line in correspondence of y=0. Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plot a scatter of the PCs from an object generated by \link{perform.PCA}.
}
\examples{
# Perform Principal Component Analyses (PCA)
pca <- perform.PCA(DEprot.object = DEprot::test.toolbox$dpo.imp)

# Plot PC1-vs-PC2 scatter
plot.PC.scatter(DEprot.PCA.object = pca,
                color.column = "condition",
                shape.column = "replicate",
                label.column = "replicate",
                plot.zero.line.y = FALSE)

# Plot PC2-vs-PC3 scatter
plot.PC.scatter(DEprot.PCA.object = pca,
                PC.x = 3,
                PC.y = 2,
                color.column = "condition",
                shape.column = "replicate",
                label.column = "replicate")


}
\seealso{
\link{perform.PCA}, \link{plot.PC.scatter.123}, \link{plot.PCoA.scatter}
}
\author{
Sebastian Gregoricchio
}
