% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PC.scatter.R
\name{plot.PC.scatter}
\alias{plot.PC.scatter}
\title{plot.PC.scatter}
\usage{
\method{plot}{PC.scatter}(
  DEprot.PCA.object,
  PC.x = 2,
  PC.y = 1,
  color.column = "column.id",
  shape.column = NULL,
  label.column = NULL,
  plot.zero.lines = TRUE
)
}
\arguments{
\item{DEprot.PCA.object}{An object of class \code{DEprot.PCA.object}, as generated by \link{perform.PCA}.}

\item{PC.x}{Number indicating which Principal Component (PC) display on the x-axis. Default: \code{2} (PC2).}

\item{PC.y}{Number indicating which Principal Component (PC) display on the y-axis. Default: \code{1} (PC1).}

\item{color.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot colors. Default: \code{"column.id"} (each sample a color).}

\item{shape.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the \code{metadata} to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{plot.zero.lines}{Logical value to indicate whether to plot two gray dashed lines in correspondence of y=0 and x=0. Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plot a scatter of the PCs from an object generated by \link{perform.PCA}.
}
\seealso{
\link{perform.PCA}
}
\author{
Sebastian Gregoricchio
}
