% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PCoA.cumulative.R
\name{plot.PCoA.cumulative}
\alias{plot.PCoA.cumulative}
\title{plot.PCoA.cumulative}
\usage{
plot.PCoA.cumulative(
  DEprot.PCoA.object,
  bar.color = "steelblue",
  line.color = "navyblue",
  title = NULL,
  broken.stick = TRUE
)
}
\arguments{
\item{DEprot.PCoA.object}{An object of class \code{DEprot.PCoA}, as generated by \link{perform.PCoA}.}

\item{bar.color}{String indicating the color to use for the bar fill. Default: \code{"steelblue"}.}

\item{line.color}{String indicating the color to use for the line and the dots of the cumulative curve. Default: \code{"navyblue"}.}

\item{title}{String indicating the title of the plot (markdown annotation supported). Default: \code{NULL}.}

\item{broken.stick}{Logical value indicating whether the broken-stick expectation should be displayed as a dashed gray line. Default: \code{TRUE}.}
}
\value{
A barplot in ggplot format.
}
\description{
Function to plot the cumulative variance of all the principal coordinates of a PCoA.
}
\details{
Only the principal coordinates corresponding to a positive eigenvalue are displayed, and the proportions of
variance are computed using the positive eigenvalues as denominator (see \link{perform.PCoA}).\cr
The broken-stick expectation provides a criterion for the axis retention: the axes whose proportion of variance
exceeds the corresponding broken-stick value are usually considered non-trivial.
}
\examples{
# Compute the sample correlations and the PCoA
correlation <- plot.correlation.heatmap(DEprot.object = DEprot::test.toolbox$dpo.imp,
                                        correlation.method = "spearman")

pcoa <- perform.PCoA(DEprot.correlation.object = correlation)

# Plot cumulative variance barplot
plot.PCoA.cumulative(pcoa)

}
\seealso{
\link{perform.PCoA}, \link{plot.PC.cumulative}
}
\author{
Sebastian Gregoricchio
}
