% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PCoA.scatter.123.R
\name{plot.PCoA.scatter.123}
\alias{plot.PCoA.scatter.123}
\title{plot.PCoA.scatter.123}
\usage{
plot.PCoA.scatter.123(
  DEprot.PCoA.object,
  color.column = "column.id",
  shape.column = NULL,
  label.column = NULL,
  dot.colors = NULL,
  title = NULL,
  plot.zero.line.y.12 = TRUE,
  plot.zero.line.x.12 = TRUE,
  plot.zero.line.y.23 = TRUE,
  plot.zero.line.x.23 = TRUE
)
}
\arguments{
\item{DEprot.PCoA.object}{An object of class \code{DEprot.PCoA}, as generated by \link{perform.PCoA}.}

\item{color.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot colors. Default: \code{"column.id"} (each sample a color).}

\item{shape.column}{String indicating the name of the column in the \code{metadata} to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the \code{metadata} to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{dot.colors}{Color-vector indicating the colors to use for the points in the plot. If \code{NULL} (default) or the number of colors is lower than the required, automatic colors will be assign using the \code{rainbow} function. Default: \code{NULL}.}

\item{title}{String indicating a title to add to the plot (markdown syntax supported). Default: \code{NULL} (no title).}

\item{plot.zero.line.y.12}{Logical value to indicate whether to plot a gray dashed line in correspondence of y=0 in the PCo1-vs-PCo2 plot. Default: \code{TRUE}.}

\item{plot.zero.line.x.12}{Logical value to indicate whether to plot a gray dashed line in correspondence of x=0 in the PCo1-vs-PCo2 plot. Default: \code{TRUE}.}

\item{plot.zero.line.y.23}{Logical value to indicate whether to plot a gray dashed line in correspondence of y=0 in the PCo2-vs-PCo3 plot. Default: \code{TRUE}.}

\item{plot.zero.line.x.23}{Logical value to indicate whether to plot a gray dashed line in correspondence of x=0 in the PCo2-vs-PCo3 plot. Default: \code{TRUE}.}
}
\value{
A patchwork object result of a combination of two ggplot objects.
}
\description{
Plot combined scatter of the principal coordinates 1-2,2-3 from an object generated by \link{perform.PCoA}.
}
\examples{
# Compute the sample correlations and the PCoA
correlation <- plot.correlation.heatmap(DEprot.object = DEprot::test.toolbox$dpo.imp,
                                        correlation.method = "spearman")

pcoa <- perform.PCoA(DEprot.correlation.object = correlation)

# Visualize the PCoA
plot.PCoA.scatter.123(DEprot.PCoA.object = pcoa,
                      color.column = "condition",
                      shape.column = "replicate",
                      label.column = "replicate")

}
\seealso{
\link{perform.PCoA}, \link{plot.PCoA.scatter}, \link{plot.PC.scatter.123}
}
\author{
Sebastian Gregoricchio
}
