% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.auroc}
\alias{plot.sPLSDA.auroc}
\title{plot.sPLSDA.auroc}
\usage{
plot.sPLSDA.auroc(
  DEprot.sPLSDA.object,
  component = NULL,
  point.colors = NULL,
  show.sd = TRUE,
  title = "**Cross-validated AUC**"
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{component}{Numeric value (or vector) indicating the component(s) to display. Default: \code{NULL} (all the components).}

\item{point.colors}{Color-vector used for the components. Default: \code{NULL} (automatic colors).}

\item{show.sd}{Logical value indicating whether the standard deviation over the repeats should be drawn as an error bar. Default: \code{TRUE}.}

\item{title}{String indicating the title of the plot (markdown annotation supported). Default: \code{"**Cross-validated AUC**"}.}
}
\value{
A ggplot object.
}
\description{
Plots the cross-validated area under the ROC curve (AUC) of each class against all the others, component by component.
}
\details{
This is the AUC computed by \code{mixOmics::perf} over the resampling, and not a ROC curve drawn on the samples used to fit the model: the latter would be optimistic by construction, since a supervised model is fitted precisely to rank those samples correctly. The dashed line at 0.5 is the value of a coin toss.

The comparisons are one class against all the others pooled, which is how mixOmics reports them for a multi-class design.
}
\examples{
\donttest{
# the performances are estimated by resampling the samples
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = TRUE,
                         folds = 3,
                         nrepeat = 1)

plot.sPLSDA.auroc(DEprot.sPLSDA.object = splsda)
}

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.performance}
}
\author{
Sebastian Gregoricchio
}
