% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.scatter.123}
\alias{plot.sPLSDA.scatter.123}
\title{plot.sPLSDA.scatter.123}
\usage{
plot.sPLSDA.scatter.123(
  DEprot.sPLSDA.object,
  color.column = NULL,
  shape.column = NULL,
  label.column = NULL,
  dot.colors = NULL,
  ellipse = TRUE,
  ellipse.level = 0.95,
  title = NULL
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{color.column}{String indicating the name of the column in the metadata to use as factor for the dot colors. Default: \code{NULL} (the column used to fit the model).}

\item{shape.column}{String indicating the name of the column in the metadata to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the metadata to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{dot.colors}{Color-vector to use for the points. Default: \code{NULL} (automatic colors).}

\item{ellipse}{Logical value indicating whether a confidence ellipse should be drawn around each group. Default: \code{TRUE}.}

\item{ellipse.level}{Numeric value (0-1) indicating the confidence level of the ellipses. Default: \code{0.95}.}

\item{title}{String indicating the title of the combined plot (markdown annotation supported). Default: \code{NULL}.}
}
\value{
A patchwork object.
}
\description{
Combines the component1-vs-component2 and component3-vs-component2 scatters of a sPLS-DA.
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

plot.sPLSDA.scatter.123(DEprot.sPLSDA.object = splsda,
                        shape.column = "replicate")

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.scatter}
}
\author{
Sebastian Gregoricchio
}
