% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.scatter}
\alias{plot.sPLSDA.scatter}
\title{plot.sPLSDA.scatter}
\usage{
plot.sPLSDA.scatter(
  DEprot.sPLSDA.object,
  comp.x = 1,
  comp.y = 2,
  color.column = NULL,
  shape.column = NULL,
  label.column = NULL,
  ellipse = TRUE,
  ellipse.level = 0.95,
  plot.zero.line.x = TRUE,
  plot.zero.line.y = TRUE,
  point.size = 3
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{comp.x}{Number indicating which component to display on the x-axis. Default: \code{1}.}

\item{comp.y}{Number indicating which component to display on the y-axis. Default: \code{2}.}

\item{color.column}{String indicating the name of the column in the metadata to use as factor for the dot colors. Default: \code{NULL} (the column used to fit the model).}

\item{shape.column}{String indicating the name of the column in the metadata to use as factor for the dot shapes. Default: \code{NULL} (all dots).}

\item{label.column}{String indicating the name of the column in the metadata to use as label of the dots. Default: \code{NULL} (no labeling).}

\item{ellipse}{Logical value indicating whether a confidence ellipse should be drawn around each group of the \code{color.column}. Default: \code{TRUE}.}

\item{ellipse.level}{Numeric value (0-1) indicating the confidence level of the ellipses. Default: \code{0.95}.}

\item{plot.zero.line.x}{Logical value to indicate whether to plot a gray dashed line in correspondence of x=0. Default: \code{TRUE}.}

\item{plot.zero.line.y}{Logical value to indicate whether to plot a gray dashed line in correspondence of y=0. Default: \code{TRUE}.}

\item{point.size}{Numeric value indicating the size of the dots. Default: \code{3}.}
}
\value{
A ggplot object.
}
\description{
Plots the scatter of two components of a sPLS-DA, as generated by \link{perform.sPLSDA}.
}
\details{
The ellipses are a reading aid, not a test: a supervised model is fitted to separate the classes, so on the training samples the groups will look tidy even when the cross-validated error rate is poor. The error rates stored in the \code{performance} slot are what says whether the separation holds (\link{plot.sPLSDA.performance}).
}
\examples{
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         keepX = 5,
                         validate = FALSE)

plot.sPLSDA.scatter(DEprot.sPLSDA.object = splsda,
                    comp.x = 1,
                    comp.y = 2,
                    shape.column = "replicate")

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.scatter.123}, \link{plot.sPLSDA.biplot}
}
\author{
Sebastian Gregoricchio
}
