% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sPLS.DA_plots.R
\name{plot.sPLSDA.tuning}
\alias{plot.sPLSDA.tuning}
\title{plot.sPLSDA.tuning}
\usage{
plot.sPLSDA.tuning(
  DEprot.sPLSDA.object,
  component = NULL,
  line.colors = NULL,
  show.sd = TRUE,
  title = "**Tuning of keepX**"
)
}
\arguments{
\item{DEprot.sPLSDA.object}{An object of class \code{DEprot.sPLSDA}.}

\item{component}{Numeric value (or vector) indicating the component(s) to display. Default: \code{NULL} (all the components).}

\item{line.colors}{Color-vector used for the components. Default: \code{NULL} (automatic colors).}

\item{show.sd}{Logical value indicating whether the standard deviation over the repeats should be drawn as a ribbon. Default: \code{TRUE}.}

\item{title}{String indicating the title of the plot (markdown annotation supported). Default: \code{"**Tuning of keepX**"}.}
}
\value{
A ggplot object.
}
\description{
Plots the classification error obtained over the grid of \code{keepX} values tested during the tuning of a sPLS-DA, and marks the value retained on each component.
}
\details{
The curves are usually flat over a wide range of values: many different numbers of proteins classify the samples equally well, and the minimum picked by the tuning is one point of a plateau rather than an optimum. Reading the plot is worth the time, because a value sitting at the left edge of that plateau gives a shorter and more interpretable protein list at no cost in error rate.
}
\examples{
\donttest{
# the tuning runs only when 'keepX' is left to NULL
splsda <- perform.sPLSDA(DEprot.object = DEprot::test.toolbox$dpo.imp,
                         group.column = "condition",
                         test.keepX = c(5, 10, 20),
                         validate = FALSE,
                         folds = 3,
                         nrepeat = 1)

plot.sPLSDA.tuning(DEprot.sPLSDA.object = splsda)
}

}
\seealso{
\link{perform.sPLSDA}, \link{plot.sPLSDA.performance}
}
\author{
Sebastian Gregoricchio
}
