% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.missing.values.R
\name{randomize.missing.values}
\alias{randomize.missing.values}
\title{randomize.missing.values}
\usage{
randomize.missing.values(
  DEprot.object,
  group.column,
  percentage.missing = 100,
  tail.percentage = 3,
  seed = floor(runif(n = 1, min = 0, max = 50000)),
  verbose = TRUE
)
}
\arguments{
\item{DEprot.object}{A \code{DEprot object}, as generated by \link{load.counts}.}

\item{group.column}{String indicating a column for the metadata table which indicates the IDs of the conditions within which the missing values should be tested.}

\item{percentage.missing}{Numeric value between 0 and 100 (included) indicated the minimal percentage of missing values per group required to apply the imputation. For instance, if \code{percentage.missing = 75} and the sample group contains 4 samples, imputation will be applied when at least 3 out of 4 values are missing. Default: \code{100} (all values are missing in the group).}

\item{tail.percentage}{Numeric value between 0 and 100 (excluded) indicating the threshold percentage of the distribution from which random values should be sampled. Default: \code{3} (percent).}

\item{seed}{Numeric value value indicating the random seed used for the imputation. By default a random integer between 0 and 50000 is selected (if \code{verbose = TRUE} a message is printed indicating the seed used).}

\item{verbose}{Logical value to define whether the message indicating which random seed was used should be printed. Default: \code{TRUE}.}
}
\value{
A \code{DEprot} object in which values for proteins which display a number of NAs above a user-defined threshold have been substituted in the normalized counts table with bottom-distribution random values.
}
\description{
Function that allows to assign random values (from the bottom-X percent of the LFQ values distribution) to proteins that show missing values for all the replicates of a specific condition.
}
