% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.config}
\alias{sample.config}
\title{Example of metadata table}
\format{
A data frame with 12 rows and 6 columns:
\describe{
  \item{\code{column.id}}{IDs of columns in the 'unimputed.counts'}
  \item{\code{sample.id}}{Actual IDs of the samples}
  \item{\code{cell}}{cell line ID}
  \item{\code{condition}}{Culture and treatment conditions}
  \item{\code{combined.id}}{ID combining cell and condition columns}
  \item{\code{replicate}}{biological replicate ID}
}
}
\source{
Simulated data
}
\usage{
sample.config
}
\description{
Dummy table for metadata corresponding to the sample configuration of \link{unimputed.counts}.
}
\keyword{datasets}
