% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse.R
\name{show,DEprot.timecourse-method}
\alias{show,DEprot.timecourse-method}
\title{show method for DEprot.timecourse objects}
\usage{
\S4method{show}{DEprot.timecourse}(object)
}
\arguments{
\item{object}{An object of class \code{DEprot.timecourse}.}
}
\value{
Prints a summary of the time-course analysis.
}
\description{
show method for DEprot.timecourse objects
}
