% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,DEprot.timecourse.enrichment-method}
\alias{show,DEprot.timecourse.enrichment-method}
\title{DEprot.timecourse.enrichment show-method}
\usage{
\S4method{show}{DEprot.timecourse.enrichment}(object)
}
\arguments{
\item{object}{An object of class \code{DEprot.timecourse.enrichment}.}
}
\value{
Prints a summary of the enrichment analyses.
}
\description{
DEprot.timecourse.enrichment show-method
}
