% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.analyses-method}
\alias{summary,DEprot.analyses-method}
\title{DEprot.analyses summary-method}
\usage{
\S4method{summary}{DEprot.analyses}(object)
}
\arguments{
\item{object}{Object of class \code{DEprot.analyses}}
}
\description{
DEprot.analyses summary-method
}
