% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{summary,DEprot.timecourse-method}
\alias{summary,DEprot.timecourse-method}
\title{summary method for DEprot.timecourse objects}
\usage{
\S4method{summary}{DEprot.timecourse}(object, ...)
}
\arguments{
\item{object}{An object of class \code{DEprot.timecourse}.}

\item{...}{Not used.}
}
\value{
A data.frame summarizing each cluster: size, dominant shape, median amplitude, median time to half-maximum, median peak time and best-ranked protein. The clusters are returned sorted by their median t.half, meaning in the order in which they respond.
}
\description{
summary method for DEprot.timecourse objects
}
